User-defined summary statistics

NeuralEstimators.samplecorrelation

NeuralEstimators.samplecovariance

NeuralEstimators.samplesize

The following functions correspond to summary statistics that are often useful as user-defined summary statistics in DeepSet objects.

NeuralEstimators.samplesize — Function

samplesize(Z::AbstractArray)

Computes the sample size of a set of independent realisations Z.

Note that this function is a wrapper around numberreplicates, but this function returns the number of replicates as the eltype of Z, rather than as an integer.

NeuralEstimators.samplecorrelation — Function

samplecorrelation(Z::AbstractArray)

Computes the sample correlation matrix, R̂, and returns the vectorised strict lower triangle of R̂.

Examples

# 5 independent replicates of a 3-dimensional vector
z = rand(3, 5)
samplecorrelation(z)

NeuralEstimators.samplecovariance — Function

samplecovariance(Z::AbstractArray)

Computes the sample covariance matrix, Σ̂, and returns the vectorised lower triangle of Σ̂.

Examples

# 5 independent replicates of a 3-dimensional vector
z = rand(3, 5)
samplecovariance(z)