Miscellaneous

Core

These functions can appear during the core workflow, and may need to be overloaded in some applications.

NeuralEstimators.subsetparameters — Function

subsetparameters(parameters::M, indices) where {M <: AbstractMatrix}
subsetparameters(parameters::P, indices) where {P <: ParameterConfigurations}

Subset parameters using a collection of indices.

Arrays in parameters::P with last dimension equal in size to the number of parameter configurations, K, are also subsetted (over their last dimension) using indices. All other fields are left unchanged. To modify this default behaviour, overload subsetparameters.

NeuralEstimators.numberreplicates — Function

numberofreplicates(Z)

Generic function that returns the number of replicates in a given object. Default implementations are provided for commonly used data formats, namely, data stored as an Array or as a GNNGraph.

NeuralEstimators.subsetdata — Function

Generic function for subsetting replicates from a data set. Default methods are:

subsetdata(Z::A, m) where {A <: AbstractArray{T, N}} where {T, N}
subsetdata(Z::G, m) where {G <: AbstractGraph}

Note that subsetdata is slow for graphical data, and one should consider using a method of train that does not require the data to be subsetted. Use numberreplicates to check that the training and validation data sets are equally replicated, which prevents the invocation of subsetdata. Note also that subsetdata only applies to vectors of batched graphs.

If the user is working with data that is not covered by the default methods, simply overload subsetdata with the appropriate type for Z.

Examples

using NeuralEstimators
using GraphNeuralNetworks
using Flux: batch

n = 5  # number of observations in each realisation
m = 6  # number of replicates for each parameter vector
d = 1  # dimension of the response variable
K = 2  # number of parameter vectors

# Array data
Z = [rand(n, d, m) for k ∈ 1:K]
subsetdata(Z, 1:3) # extract first 3 replicates for each parameter vector

# Graphical data
e = 8 # number of edges
Z = [batch([rand_graph(n, e, ndata = rand(d, n)) for _ ∈ 1:m]) for k ∈ 1:K]
subsetdata(Z, 1:3) # extract first 3 replicates for each parameter vector

Activation functions

These layers can be used at the end of an architecture to ensure that the neural estimator provides valid parameters.

NeuralEstimators.Compress — Type

Compress(a, b)

Layer to compress the output to be within a and b, where each element of a is less than the corresponding element of b.

Examples

using NeuralEstimators
using Flux

p = 3
a = [0.1, -1, 2]
b = [0.9, 1, 3]
n = 20
K = 10
θ̂ = Chain(Dense(n, p), Compress(a, b))
Z = rand(n, K)
θ̂(Z)

NeuralEstimators.CholeskyCovariance — Type

CholeskyCovariance(d)

Layer for constructing the parameters of the lower Cholesky factor associated with an unconstrained d×d covariance matrix.

The layer transforms a Matrix with d(d+1)÷2 rows into a Matrix of the same dimension, but with d rows constrained to be positive (corresponding to the diagonal elements of the Cholesky factor) and the remaining rows unconstrained.

The ordering of the transformed Matrix aligns with Julia's column-major ordering. For example, when modelling the Cholesky factor,

\[\begin{bmatrix} L₁₁ & & \\ L₂₁ & L₂₂ & \\ L₃₁ & L₃₂ & L₃₃ \\ \end{bmatrix},\]

the rows of the matrix returned by a CholeskyCovariance layer will be ordered as

\[L₁₁, L₂₁, L₃₁, L₂₂, L₃₂, L₃₃,\]

which means that the output can easily be transformed into the implied Cholesky factors using vectotril.

Examples

using NeuralEstimators

d = 4
p = d*(d+1)÷2
θ = randn(p, 50)
l = CholeskyCovariance(d)
θ = l(θ)                              # returns matrix (used for Flux networks)
L = [vectotril(y) for y ∈ eachcol(θ)] # convert matrix to Cholesky factors

NeuralEstimators.CovarianceMatrix — Type

CovarianceMatrix(d)

Layer for constructing the parameters of an unconstrained d×d covariance matrix.

The layer transforms a Matrix with d(d+1)÷2 rows into a Matrix of the same dimension.

Internally, it uses a CholeskyCovariance layer to construct a valid Cholesky factor 𝐋, and then extracts the lower triangle from the positive-definite covariance matrix 𝚺 = 𝐋𝐋'. The lower triangle is extracted and vectorised in line with Julia's column-major ordering. For example, when modelling the covariance matrix,

\[\begin{bmatrix} Σ₁₁ & Σ₁₂ & Σ₁₃ \\ Σ₂₁ & Σ₂₂ & Σ₂₃ \\ Σ₃₁ & Σ₃₂ & Σ₃₃ \\ \end{bmatrix},\]

the rows of the matrix returned by a CovarianceMatrix layer will be ordered as

\[Σ₁₁, Σ₂₁, Σ₃₁, Σ₂₂, Σ₃₂, Σ₃₃,\]

which means that the output can easily be transformed into the implied covariance matrices using vectotril and Symmetric.

Examples

using NeuralEstimators
using LinearAlgebra

d = 4
p = d*(d+1)÷2
θ = randn(p, 50)

l = CovarianceMatrix(d)
θ = l(θ)
Σ = [Symmetric(cpu(vectotril(y)), :L) for y ∈ eachcol(θ)]

NeuralEstimators.CorrelationMatrix — Type

CorrelationMatrix(d)

Layer for constructing the parameters of an unconstrained d×d correlation matrix.

The layer transforms a Matrix with d(d-1)÷2 rows into a Matrix with the same dimension.

Internally, the layers uses the algorithm described here and here to construct a valid Cholesky factor 𝐋, and then extracts the strict lower triangle from the positive-definite correlation matrix 𝐑 = 𝐋𝐋'. The strict lower triangle is extracted and vectorised in line with Julia's column-major ordering. For example, when modelling the correlation matrix,

\[\begin{bmatrix} 1 & R₁₂ & R₁₃ \\ R₂₁ & 1 & R₂₃\\ R₃₁ & R₃₂ & 1\\ \end{bmatrix},\]

the rows of the matrix returned by a CorrelationMatrix layer will be ordered as

\[R₂₁, R₃₁, R₃₂,\]

which means that the output can easily be transformed into the implied correlation matrices using the strict variant of vectotril and Symmetric.

Examples

using NeuralEstimators
using LinearAlgebra

d = 4
p = d*(d-1)÷2
l = CorrelationMatrix(d)
θ = randn(p, 50)

# returns a matrix of parameters
θ = l(θ)

# convert matrix of parameters to implied correlation matrices
R = map(eachcol(θ)) do y
	R = Symmetric(cpu(vectotril(y, strict = true)), :L)
	R[diagind(R)] .= 1
	R
end

NeuralEstimators.SplitApply — Type

SplitApply(layers, indices)

Splits an array into multiple sub-arrays by subsetting the rows using the collection of indices, and then applies each layer in layers to the corresponding sub-array.

Specifically, for each i = 1, …, $n$, with $n$ the number of layers, SplitApply(x) performs layers[i](x[indices[i], :]), and then vertically concatenates the resulting transformed arrays.

Examples

using NeuralEstimators

d = 4
K = 50
p₁ = 2          # number of non-covariance matrix parameters
p₂ = d*(d+1)÷2  # number of covariance matrix parameters
p = p₁ + p₂

a = [0.1, 4]
b = [0.9, 9]
l₁ = Compress(a, b)
l₂ = CovarianceMatrix(d)
l = SplitApply([l₁, l₂], [1:p₁, p₁+1:p])

θ = randn(p, K)
l(θ)

Utility functions

NeuralEstimators.vectotril — Function

vectotril(v; strict = false)
vectotriu(v; strict = false)

Converts a vector v of length $d(d+1)÷2$ (a triangular number) into a $d × d$ lower or upper triangular matrix.

If strict = true, the matrix will be strictly lower or upper triangular, that is, a $(d+1) × (d+1)$ triangular matrix with zero diagonal.

Note that the triangular matrix is constructed on the CPU, but the returned matrix will be a GPU array if v is a GPU array. Note also that the return type is not of type Triangular matrix (i.e., the zeros are materialised) since Traingular matrices are not always compatible with other GPU operations.

Examples

using NeuralEstimators

d = 4
n = d*(d+1)÷2
v = collect(range(1, n))
vectotril(v)
vectotriu(v)
vectotril(v; strict = true)
vectotriu(v; strict = true)

NeuralEstimators.containertype — Function

containertype(A::Type)
containertype(::Type{A}) where A <: SubArray
containertype(a::A) where A

Returns the container type of its argument.

If given a SubArray, returns the container type of the parent array.

Examples

a = rand(3, 4)
containertype(a)
containertype(typeof(a))
[containertype(x) for x ∈ eachcol(a)]

NeuralEstimators.loadbestweights — Function

loadbestweights(path::String)

Given a path to a training run containing neural networks saved with names "network_epochx.bson" and an object saved as "loss_per_epoch.bson", returns the weights of the best network (measured by validation loss).

NeuralEstimators.stackarrays — Function

stackarrays(v::V; merge = true) where {V <: AbstractVector{A}} where {A <: AbstractArray{T, N}} where {T, N}

Stack a vector of arrays v along the last dimension of each array, optionally merging the final dimension of the stacked array.

The arrays must be of the same size for the first N-1 dimensions. However, if merge = true, the size of the final dimension can vary.

Examples

# Vector containing arrays of the same size:
Z = [rand(2, 3, m) for m ∈ (1, 1)];
stackarrays(Z)
stackarrays(Z, merge = false)

# Vector containing arrays with differing final dimension size:
Z = [rand(2, 3, m) for m ∈ (1, 2)];
stackarrays(Z)

NeuralEstimators.expandgrid — Function

expandgrid(xs, ys)

Same as expand.grid() in R, but currently caters for two dimensions only.