Core functions

Parameters

NeuralEstimators.ParameterConfigurations — Type

ParameterConfigurations

An abstract supertype for storing parameters θ and any intermediate objects needed for data simulation with simulate.

NeuralEstimators.subsetparameters — Function

subsetparameters(parameters::Parameters, indices) where {Parameters <: ParameterConfigurations}

Subset parameters using a collection of indices.

The default method assumes that each field of parameters is an array. If the last dimension of the array has size equal to the number of parameter configurations, K, then the array is subsetted over its last dimension using indices; otherwise, the field is returned unchanged. If this default does not cover your use case, define an appropriate subsetting method by overloading subsetparameters after running import NeuralEstimators: subsetparameters.

Simulation

NeuralEstimators.simulate — Function

simulate(parameters::P, ξ, m::Integer, num_rep::Integer) where {P <: ParameterConfigurations}

Generic method that simulates num_rep sets of sets of m independent replicates for each parameter configuration by calling simulate(parameters, ξ, m).

Deep Set representation

Vanilla Deep Set

NeuralEstimators.DeepSet — Type

DeepSet(ψ, ϕ, agg)

Implementation of the Deep Set framework, where ψ and ϕ are neural networks (e.g., Flux networks) and agg is a symmetric function that pools data over the last dimension (the replicates/batch dimension) of an array.

DeepSet objects are applied to AbstractVectors of AbstractArrays, where each array is associated with one parameter vector.

Examples

n = 10 # observations in each realisation
p = 5  # number of parameters in the statistical model
w = 32 # width of each layer
ψ = Chain(Dense(n, w, relu), Dense(w, w, relu));
ϕ = Chain(Dense(w, w, relu), Dense(w, p));
agg(X) = sum(X, dims = ndims(X))
θ̂  = DeepSet(ψ, ϕ, agg)

# A single set of m=3 realisations:
Z = [rand(n, 1, 3)];
θ̂ (Z)

# Two sets each containing m=3 realisations:
Z = [rand(n, 1, m) for m ∈ (3, 3)];
θ̂ (Z)

# Two sets respectivaly containing m=3 and m=4 realisations:
Z = [rand(n, 1, m) for m ∈ (3, 4)];
θ̂ (Z)

NeuralEstimators.DeepSet — Method

DeepSet(ψ, ϕ; aggregation::String = "mean")

Convenient constructor for a DeepSet object with agg equal to the "mean", "sum", or "logsumexp" function.

Deep Set with expert summary statistics

NeuralEstimators.DeepSetExpert — Type

DeepSetExpert(ψ, ϕ, S, agg)

Implementation of the Deep Set framework with ψ and ϕ neural networks, agg a symmetric function that pools data over the last dimension of an array, and S a vector of real-valued functions that serve as expert summary statistics.

The dimension of the domain of ϕ should be qₜ + qₛ, where qₜ is the range of ϕ and qₛ is the dimension of S, that is, length(S). DeepSetExpert objects are applied to AbstractVectors of AbstractArrays, where each array is associated with one parameter vector. The functions ψ and S both act on these arrays individually (i.e., they are broadcasted over the AbstractVector).

NeuralEstimators.DeepSetExpert — Method

DeepSetExpert(deepset::DeepSet, ϕ, S)

DeepSetExpert constructor with the aggregation function agg and inner neural network ψ inherited from deepset.

Note that we cannot inherit the outer network, ϕ, since DeepSetExpert objects require the dimension of the domain of ϕ to be qₜ + qₛ.

NeuralEstimators.DeepSetExpert — Method

DeepSetExpert(ψ, ϕ, S; aggregation::String = "mean")

DeepSetExpert constructor with agg equal to the "mean", "sum", or "logsumexp" function.

NeuralEstimators.samplesize — Function

samplesize(Z::A) where {A <: AbstractArray{T, N}} where {T, N}

Computes the sample size m for a set of independent realisations Z, useful as an expert summary statistic in DeepSetExpert objects.

Piecewise Deep Set neural estimators

NeuralEstimators.DeepSetPiecewise — Type

DeepSetPiecewise(estimators, m_cutoffs)

Given an arbitrary number of estimators, creates a piecewise neural estimator based on the sample size cut offs, m_cutoffs, which should contain one element fewer than the number of estimators.

Examples

Suppose that we have two neural estimators, θ̂₁ and θ̂₂, taking the following arbitrary forms:

n = 10
p = 5
w = 32

ψ₁ = Chain(Dense(n, w, relu), Dense(w, w, relu));
ϕ₁ = Chain(Dense(w, w, relu), Dense(w, p));
θ̂₁ = DeepSet(ψ₁, ϕ₁)

ψ₂ = Chain(Dense(n, w, relu), Dense(w, w, relu), Dense(w, w, relu));
ϕ₂ = Chain(Dense(w, w, relu), Dense(w, w, relu), Dense(w, p));
θ̂₂ = DeepSet(ψ₂, ϕ₂)

Further suppose that we've trained θ̂₁ for small sample sizes (e.g., m ≤ 30) and θ̂₂ for moderate-to-large sample sizes (e.g., m > 30). Then we construct a piecewise Deep Set object with a cut-off sample size of 30 which dispatches θ̂₁ if m ≤ 30 and θ̂₂ if m > 30:

θ̂ = DeepSetPiecewise((θ̂₁, θ̂₂), (30,))
Z = [rand(Float32, n, 1, m) for m ∈ (10, 50)]
θ̂(Z)

Training

There are two training methods. For both methods, the validation parameters and validation data are held fixed so that the validation risk is interpretable. There are a number of practical considerations to keep in mind: In particular, see Balancing time and memory complexity during training.

NeuralEstimators.train — Function

train(θ̂, ξ, P; <keyword args>) where {P <: ParameterConfigurations}

Train the neural estimator θ̂ by providing the invariant model information ξ needed for the constructor P to automatically sample the sets of training and validation parameters.

Keyword arguments common to both train methods:

m: sample sizes (either an Integer or a collection of Integers).

batchsize::Integer = 32

epochs::Integer = 100: the maximum number of epochs used during training.

epochs_per_Z_refresh::Integer = 1: how often to refresh the training data.

loss = mae: the loss function, which should return an average loss when applied to multiple replicates.

optimiser = ADAM(1e-4)

savepath::String = "runs/": path to save the trained θ̂ and other information; if savepath is an empty string (i.e., ""), nothing is saved.

simulate_just_in_time::Bool = false: should we do "just-in-time" data simulation, which improves memory complexity at the cost of time complexity?

stopping_epochs::Integer = 10: cease training if the risk doesn't improve in stopping_epochs epochs.

use_gpu::Bool = true

verbose::Bool = true

Simulator keyword arguments only:

K::Integer = 10_000: the number of parameters in the training set; the size of the validation set is K ÷ 5.

epochs_per_θ_refresh::Integer = 1: how often to refresh the training parameters; this must be a multiple of epochs_per_Z_refresh.

train(θ̂, ξ, θ_train::P, θ_val::P; <keyword args>) where {P <: ParameterConfigurations}

Train the neural estimator θ̂ by providing the training and validation parameter sets explicitly as θ_train and θ_val, which are both held fixed during training, as well as the invariant model information ξ.

train(θ̂, θ_train::P, θ_val::P, Z_train::T, Z_val::T; <keyword args>) where {T, P <: ParameterConfigurations}

Train the neural estimator θ̂ by providing the training and validation parameter sets, θ_train and θ_val, and the training and validation data sets, Z_train and Z_val, all of which are held fixed during training.

The sample size argument m is inferred from Z_val. The training data Z_train can contain M replicates, where M is a multiple of m; the training data will then be recycled to imitate on-the-fly simulation. For example, if M = 50 and m = 10, epoch 1 uses the first 10 replicates, epoch 2 uses the second 10 replicates, and so on, until epoch 6 again uses the first 10 replicates.

Note that the elements of Z_train and Z_val should be equally replicated; that is, the size of the last dimension in each array in Z_train should be constant, and similarly for Z_val.

Estimation

NeuralEstimators.estimate — Function

estimate(estimators, ξ, parameters::P; <keyword args>) where {P <: ParameterConfigurations}

Using a collection of estimators, compute estimates from data simulated from a set of parameters with invariant information ξ.

Note that estimate() requires the user to have defined a method simulate(parameters, ξ, m::Integer).

Keyword arguments

m::Vector{Integer}: sample sizes to estimate from.

estimator_names::Vector{String}: names of the estimators (sensible default values provided).

parameter_names::Vector{String}: names of the parameters (sensible default values provided).

num_rep::Integer = 1: the number of times to replicate each parameter in parameters.

save::Vector{String}: by default, no objects are saved; however, if save is provided, four DataFrames respectively containing the true parameters θ, estimates θ̂, runtimes, and merged θ and θ̂ will be saved in the directory save[1] with file names (not extensions) suffixed by save[2].

use_ξ = false: a Bool or a collection of Bool objects with length equal to the number of estimators. Specifies whether or not the estimator uses the invariant model information, ξ: If it does, the estimator will be applied as estimator(Z, ξ).

use_gpu = true: a Bool or a collection of Bool objects with length equal to the number of estimators.

verbose::Bool = true

NeuralEstimators.Estimates — Type

Estimates(θ, θ̂, runtime)

A set of true parameters θ, corresponding estimates θ̂, and the runtime to obtain θ̂, as returned by a call to estimate.

Base.merge — Method

merge(estimates::Estimates)

Merge estimates into a single long-form DataFrame containing the true parameters and the corresponding estimates.

Bootstrapping

Note that all bootstrapping functions are currently implemented for a single parameter configuration only.

NeuralEstimators.parametricbootstrap — Function

parametricbootstrap(θ̂, parameters::P, ξ, m::Integer; B::Integer = 100, use_gpu::Bool = true) where {P <: ParameterConfigurations}

Returns B parameteric bootstrap samples of an estimator θ̂ as a p × B matrix, where p is the number of parameters in the statistical model, based on data sets of size m simulated using the invariant model information ξ and parameter configurations, parameters.

This function requires the user to have defined a method simulate(parameters::P, ξ, m::Integer).

NeuralEstimators.nonparametricbootstrap — Function

nonparametricbootstrap(θ̂, Z::AbstractArray{T, N}; B::Integer = 100, use_gpu::Bool = true)
nonparametricbootstrap(θ̂, Z::AbstractArray{T, N}, blocks; B::Integer = 100, use_gpu::Bool = true)

Returns B non-parametric bootstrap samples of an estimator θ̂ as a p × B matrix, where p is the number of parameters in the statistical model.

The argument blocks caters for block bootstrapping, and should be an integer vector specifying the block for each replicate. For example, if we have 5 replicates with the first two replicates corresponding to block 1 and the remaining replicates corresponding to block 2, then blocks should be [1, 1, 2, 2, 2]. The resampling algorithm tries to produce resampled data sets of a similar size to the original data, but this can only be achieved exactly if the blocks are the same length.