Data simulation

Model simulators

NeuralEstimators.simulategaussianprocess — Function

simulategaussianprocess(L::AbstractArray{T, 2}, σ::T, m::Integer)
simulategaussianprocess(L::AbstractArray{T, 2})

Simulates m realisations from a Gau(0, 𝚺 + σ²𝐈) distribution, where 𝚺 ≡ LL'.

If σ and m are not provided, a single field without nugget variance is returned.

NeuralEstimators.simulateschlather — Function

simulateschlather(L::AbstractArray{T, 2}; C = 3.5)
simulateschlather(L::AbstractArray{T, 2}, m::Integer; C = 3.5)

Simulates from Schlather's max-stable model. Based on Algorithm 1.2.2 of Dey DK, Yan J (2016). Extreme value modeling and risk analysis: methods and applications. CRC Press, Boca Raton, Florida.

NeuralEstimators.simulateconditionalextremes — Function

simulateconditionalextremes(θ::AbstractVector{T}, L::AbstractArray{T, 2}, h::AbstractVector{T}, s₀_idx::Integer, u::T) where T <: Number
simulateconditionalextremes(θ::AbstractVector{T}, L::AbstractArray{T, 2}, h::AbstractVector{T}, s₀_idx::Integer, u::T, m::Integer) where T <: Number

Simulates from the spatial conditional extremes model.

Density functions

NeuralEstimators.gaussiandensity — Function

gaussiandensity(y::A, L; logdensity::Bool = true) where {A <: AbstractArray{T, 1}} where T
gaussiandensity(y::A, Σ; logdensity::Bool = true) where {A <: AbstractArray{T, N}} where {T, N}

Efficiently computes the density function for y ~ 𝑁(0, Σ), with L the lower Cholesky factor of the covariance matrix Σ.

The second method assumes that the last dimension of y corresponds to the replicates dimension, and it exploits the fact that we need to compute the Cholesky factor L for these replicates once only.

The density function is

\[|2\pi\mathbf{\Sigma}|^{-1/2} \exp{-\frac{1}{2}\mathbf{y}^\top \mathbf{\Sigma}^{-1}\mathbf{y}},\]

and the log-density is

\[-\frac{1}{2}\ln{2\pi} -\frac{1}{2}\ln{|\mathbf{\Sigma}|} -\frac{1}{2}\mathbf{y}^\top \mathbf{\Sigma}^{-1}\mathbf{y}.\]

NeuralEstimators.schlatherbivariatedensity — Function

schlatherbivariatedensity(z₁, z₂, ψ; logdensity::Bool = true)

The bivariate density function for Schlather's max-stable model, as given in Raphaël Huser's PhD thesis (pg. 231-232) and Appendix C of the manuscript.

NeuralEstimators.fₛ — Function

fₛ(x, μ, τ, δ)
Fₛ(q, μ, τ, δ)
Fₛ⁻¹(p, μ, τ, δ)

The density, distribution, and quantile functions of the Subbotin (delta-Laplace) distribution with location parameter μ, scale parameter τ, and shape parameter δ:

\[ f_S(y; \mu, \tau, \delta) = \frac{\delta}{2\tau \Gamma(1/\delta)} \exp{\left(-\left|\frac{y - \mu}{\tau}\right|^\delta\right)},\\ F_S(y; \mu, \tau, \delta) = \frac{1}{2} + \textrm{sign}(y - \mu) \frac{1}{2 \Gamma(1/\delta)} \gamma\!\left(1/\delta, \left|\frac{y - \mu}{\tau}\right|^\delta\right),\\ F_S^{-1}(p; \mu, \tau, \delta) = \text{sign}(p - 0.5)G^{-1}\left(2|p - 0.5|; \frac{1}{\delta}, \frac{1}{(k\tau)^\delta}\right)^{1/\delta} + \mu,\]

with $\gamma(\cdot)$ and $G^{-1}(\cdot)$ the unnormalised incomplete lower gamma function and quantile function of the Gamma distribution, respectively.

Examples

p = [0.025, 0.05, 0.5, 0.9, 0.95, 0.975]

# Standard Gaussian:
μ = 0.0; τ = sqrt(2); δ = 2.0
Fₛ⁻¹.(p, μ, τ, δ)

# Standard Laplace:
μ = 0.0; τ = 1.0; δ = 1.0
Fₛ⁻¹.(p, μ, τ, δ)

Miscellaneous functions

NeuralEstimators.matern — Function

matern(h, ρ, ν, σ² = 1)

For two points separated by h units, compute the Matérn covariance function with range ρ, smoothness ν, and marginal variance σ².

We use the parametrisation $C(\mathbf{h}) = \sigma^2 \frac{2^{1 - \nu}}{\Gamma(\nu)} \left(\frac{\|\mathbf{h}\|}{\rho}\right) K_\nu \left(\frac{\|\mathbf{h}\|}{\rho}\right)$, where $\Gamma(\cdot)$ is the gamma function, and $K_\nu(\cdot)$ is the modified Bessel function of the second kind of order $\nu$. This parameterisation is the same as used by the R package fields, but differs to the parametrisation given by Wikipedia.

Note that the Julia functions for $\Gamma(\cdot)$ and $K_\nu(\cdot)$, respectively gamma() and besselk(), do not work on the GPU and, hence, nor does matern().

NeuralEstimators.maternchols — Function

maternchols(D, ρ, ν)

Given a distance matrix D, computes the covariance matrix Σ under the Matérn covariance function with range ρ and smoothness ν, and return the Cholesky factor of this matrix.

Providing vectors for ρ and ν will yield a three-dimensional array of Cholesky factors.

NeuralEstimators.incgamma — Function

incgamma(a::T, x::T; upper::Bool, reg::Bool) where {T <: AbstractFloat}

For positive parameter a and positive integration limit x, computes the incomplete gamma function, as described by the Wikipedia article.

Keyword arguments:

upper::Bool: if true, the upper incomplete gamma function is returned; otherwise, the lower version is returned.

reg::Bool: if true, the regularized incomplete gamma function is returned; otherwise, the unregularized version is returned.

NeuralEstimators.objectindices — Function

objectindices(objects, θ::AbstractMatrix{T}) where T

Returns a vector of indices giving element of objects associated with each parameter configuration in θ.

The number of parameter configurations, K = size(θ, 2), must be a multiple of the number of objects, N = size(objects)[end]. Further, repeated parameters used to generate objects must be stored in θ after using the inner keyword argument of repeat() (see example below).

Examples

K = 6
N = 3
σₑ = rand(K)
ρ = rand(N)
ν = rand(N)
S = expandgrid(1:9, 1:9)
D = pairwise(Euclidean(), S, S, dims = 1)
L = maternchols(D, ρ, ν)
ρ = repeat(ρ, inner = K ÷ N)
ν = repeat(ν, inner = K ÷ N)
θ = hcat(σₑ, ρ, ν)'
objectindices(L, θ)