Utility functions

NeuralEstimators.loadbestweights — Function

loadbestweights(path::String)

Given a path to a training run containing neural networks saved with names 'networkepochx.bson' and an object saved as 'lossper_epoch.bson', returns the weights of the best network (measured by validation loss).

NeuralEstimators.stackarrays — Function

stackarrays(v::V; merge::Bool = true) where {V <: AbstractVector{A}} where {A <: AbstractArray{T, N}} where {T, N}

Stack a vector of arrays v along the last dimension of each array, optionally merging the final dimension of the stacked array.

The arrays must be of the same for the first N-1 dimensions. However, if merge = true, the size of the final dimension can vary between arrays.

Examples

# Vector containing arrays of the same size:
Z = [rand(2, 3, m) for m ∈ (1, 1)];
stackarrays(Z)
stackarrays(Z, merge = false)

# Vector containing arrays with differing final dimension size:
Z = [rand(2, 3, m) for m ∈ (1, 2)];
stackarrays(Z)

NeuralEstimators.expandgrid — Function

expandgrid(xs, ys)

Same as expand.grid() in R, but currently caters for two dimensions only.