Core

This page documents the functions that are central to the workflow of NeuralEstimators. Its organisation reflects the order in which these functions appear in a standard implementation; that is, from sampling parameters from the prior distribution, to uncertainty quantification of the final estimates via bootstrapping.

Sampling parameters

Parameters sampled from the prior distribution $\Omega(\cdot)$ are stored as a $p \times K$ matrix, where $p$ is the number of parameters in the model and $K$ is the number of parameter vectors sampled from the prior distribution.

It can sometimes be helpful to wrap the parameter matrix in a user-defined type that also stores expensive intermediate objects needed for data simulated (e.g., Cholesky factors). In this case, the user-defined type should be a subtype of the abstract type ParameterConfigurations, whose only requirement is a field θ that stores the matrix of parameters. See Storing expensive intermediate objects for data simulation for further discussion.

NeuralEstimators.ParameterConfigurations — Type

ParameterConfigurations

An abstract supertype for user-defined types that store parameters and any intermediate objects needed for data simulation.

The user-defined type must have a field θ that stores the $p$ × $K$ matrix of parameters, where $p$ is the number of parameters in the model and $K$ is the number of parameter vectors sampled from the prior distribution. There are no other restrictions.

See subsetparameters for the generic function for subsetting these objects.

Examples

struct P <: ParameterConfigurations
	θ
	# other expensive intermediate objects...
end

Simulating data

NeuralEstimators facilitates neural estimation for arbitrary statistical models by having the user implicitly define the model via simulated data. The user may provide simulated data directly, or provide a function that simulates data from the model.

The data should be stored as a Vector{A}, where each element of the vector is associated with one parameter configuration, and where A depends on the architecture of the neural estimator.

Types of estimators

See also Architectures and activations functions that are often used when constructing neural estimators.

NeuralEstimators.NeuralEstimator — Type

NeuralEstimator

An abstract supertype for neural estimators.

NeuralEstimators.PointEstimator — Type

PointEstimator(arch)

A simple point estimator, that is, a mapping from the sample space to the parameter space, defined by the given architecture arch.

NeuralEstimators.IntervalEstimator — Type

IntervalEstimator(arch_lower, arch_upper)
IntervalEstimator(arch)

A neural interval estimator that jointly estimates credible intervals constructed as,

\[[l(Z), l(Z) + \mathrm{exp}(u(Z))],\]

where $l(⋅)$ and $u(⋅)$ are the neural networks arch_lower and arch_upper, both of which should transform data into $p$-dimensional vectors, where $p$ is the number of parameters in the statistical model. If only a single neural network architecture arch is provided, it will be used for both arch_lower and arch_upper.

The returned value is a matrix with $2p$ rows, where the first and second $p$ rows correspond to estimates of the lower and upper bound, respectively.

See also IntervalEstimatorCompactPrior.

Examples

using NeuralEstimators
using Flux

# Generate some toy data
n = 2   # bivariate data
m = 100 # number of independent replicates
Z = rand(n, m)

# Create an architecture
p = 3  # number of parameters in the statistical model
w = 8  # width of each layer
ψ = Chain(Dense(n, w, relu), Dense(w, w, relu));
ϕ = Chain(Dense(w, w, relu), Dense(w, p));
architecture = DeepSet(ψ, ϕ)

# Initialise the interval estimator
estimator = IntervalEstimator(architecture)

# Apply the interval estimator
estimator(Z)
interval(estimator, Z)

NeuralEstimators.IntervalEstimatorCompactPrior — Type

IntervalEstimatorCompactPrior(u, v, min_supp::Vector, max_supp::Vector)
IntervalEstimatorCompactPrior(u, v, compress::Compress)

Uses the neural networks u and v to jointly estimate credible intervals that are guaranteed to be within the support of the prior distributon. This support is defined by the $p$-dimensional vectors min_supp and max_supp (or a single $p$-dimensional object of type Compress), where $p$ is the number of parameters in the statistical model.

Given data $Z$, the intervals are constructed as

\[[g(u(Z)), g(u(Z)) + \mathrm{exp}(v(Z)))],\]

where

$u(⋅)$ and $v(⋅)$ are neural networks, both of which should transform data into $p$-dimensional vectors;

$g(⋅)$ is a logistic function that maps its input to the prior support.

Note that, in addition to ensuring that the interval remains in the prior support, this constructions also ensures that the intervals are valid (i.e., it prevents quantile crossing, in the sense that the upper bound is always greater than the lower bound).

The returned value is a matrix with $2p$ rows, where the first and second $p$ rows correspond to estimates of the lower and upper bound, respectively.

See also IntervalEstimator and Compress.

Examples

using NeuralEstimators
using Flux

# prior support
min_supp = [25, 0.5, -pi/2]
max_supp = [500, 2.5, 0]
p = length(min_supp)  # number of parameters in the statistical model

# Generate some toy data
n = 2   # bivariate data
m = 100 # number of independent replicates
Z = rand(n, m)

# Create an architecture
w = 8  # width of each layer
ψ = Chain(Dense(n, w, relu), Dense(w, w, relu));
ϕ = Chain(Dense(w, w, relu), Dense(w, p));
u = DeepSet(ψ, ϕ)
v = deepcopy(u) # use the same architecture for both u and v

# Initialise the interval estimator
estimator = IntervalEstimatorCompactPrior(u, v, min_supp, max_supp)

# Apply the interval estimator
estimator(Z)
interval(estimator, Z)

NeuralEstimators.PointIntervalEstimator — Type

PointIntervalEstimator(arch_point, arch_lower, arch_upper)
PointIntervalEstimator(arch_point, arch_bound)
PointIntervalEstimator(arch)

A neural estimator that jointly produces point estimates, θ̂(Z), where θ̂(Z) is a neural point estimator with architecture arch_point, and credible intervals constructed as,

\[[θ̂(Z) - \mathrm{exp}(l(Z)), θ̂(Z) + \mathrm{exp}(u(Z))],\]

where $l(⋅)$ and $u(⋅)$ are the neural networks arch_lower and arch_upper, both of which should transform data into $p$-dimensional vectors, where $p$ is the number of parameters in the statistical model.

If only a single neural network architecture arch is provided, it will be used for all architectures; similarly, if two architectures are provided, the second will be used for both arch_lower and arch_upper.

Internally, the point estimates, lower-bound estimates, and upper-bound estimates are concatenated, so that PointIntervalEstimator objects transform data into matrices with $3p$ rows.

Examples

using NeuralEstimators
using Flux

# Generate some toy data
n = 2   # bivariate data
m = 100 # number of independent replicates
Z = rand(n, m)

# Create an architecture
p = 3  # number of parameters in the statistical model
w = 8  # width of each layer
ψ = Chain(Dense(n, w, relu), Dense(w, w, relu));
ϕ = Chain(Dense(w, w, relu), Dense(w, p));
architecture = DeepSet(ψ, ϕ)

# Initialise the estimator
estimator = PointIntervalEstimator(architecture)

# Apply the estimator
estimator(Z)
interval(estimator, Z)

NeuralEstimators.QuantileEstimator — Type

QuantileEstimator()

Coming soon: this structure will allow for the simultaneous estimation of an arbitrary number of marginal quantiles of the posterior distribution.

NeuralEstimators.PiecewiseEstimator — Type

PiecewiseEstimator(estimators, breaks)

Creates a piecewise estimator from a collection of estimators, based on the collection of changepoints, breaks, which should contain one element fewer than the number of estimators.

Any estimator can be included in estimators, including any of the subtypes of NeuralEstimator exported with the package NeuralEstimators (e.g., PointEstimator, IntervalEstimator, etc.).

Examples

# Suppose that we've trained two neural estimators. The first, θ̂₁, is trained
# for small sample sizes (e.g., m ≤ 30), and the second, `θ̂₂`, is trained for
# moderate-to-large sample sizes (e.g., m > 30). We construct a piecewise
# estimator with a sample-size changepoint of 30, which dispatches θ̂₁ if m ≤ 30
# and θ̂₂ if m > 30.

using NeuralEstimators
using Flux

n = 2  # bivariate data
p = 3  # number of parameters in the statistical model
w = 8  # width of each layer

ψ₁ = Chain(Dense(n, w, relu), Dense(w, w, relu));
ϕ₁ = Chain(Dense(w, w, relu), Dense(w, p));
θ̂₁ = DeepSet(ψ₁, ϕ₁)

ψ₂ = Chain(Dense(n, w, relu), Dense(w, w, relu));
ϕ₂ = Chain(Dense(w, w, relu), Dense(w, p));
θ̂₂ = DeepSet(ψ₂, ϕ₂)

θ̂ = PiecewiseEstimator([θ̂₁, θ̂₂], [30])
Z = [rand(n, 1, m) for m ∈ (10, 50)]
θ̂(Z)

Training

The function train is used to train a single neural estimator, while the wrapper function trainx is useful for training multiple neural estimators over a range of sample sizes, making using of the technique known as pre-training.

NeuralEstimators.train — Function

train(θ̂, sampler, simulator; )
train(θ̂, θ_train::P, θ_val::P, simulator; ) where {P <: Union{AbstractMatrix, ParameterConfigurations}}
train(θ̂, θ_train::P, θ_val::P, Z_train::T, Z_val::T; ) where {T, P <: Union{AbstractMatrix, ParameterConfigurations}}

Train a neural estimator with architecture θ̂.

The methods cater for different forms of on-the-fly simulation. The method that takes functions sampler and simulator for sampling parameters and simulating data, respectively, allows for both the parameters and the data to be simulated on-the-fly. Note that simulator is called as simulator(θ, m), where θ is a set of parameters and m is the sample size (see keyword arguments below). If provided with specific instances of parameters (θ_train and θ_val) or data (Z_train and Z_val), they will be held fixed during training.

In all methods, the validation set is held fixed to reduce noise when evaluating the validation risk function, which is used to monitor the performance of the estimator during training.

If the number of replicates in Z_train is a multiple of the number of replicates for each element of Z_val, the training data will be recycled throughout training. For example, if each element of Z_train consists of 50 replicates, and each element of Z_val consists of 10 replicates, the first epoch uses the first 10 replicates in Z_train, the second epoch uses the next 10 replicates, and so on, until the sixth epoch again uses the first 10 replicates. Note that this requires the data to be subsettable with the function subsetdata.

Keyword arguments

Arguments common to all methods:

loss = mae: the loss function, which should return the average loss when applied to multiple replicates.

epochs::Integer = 100

batchsize::Integer = 32

optimiser = ADAM(1e-4)

savepath::String = "": path to save the neural-network weights during training (as bson files) and other information, such as the risk function per epoch (the risk function evaluated over the training and validation sets are saved in the first and second columns of loss_per_epoch.csv). If savepath is an empty string (default), nothing is saved.

stopping_epochs::Integer = 5: cease training if the risk doesn't improve in this number of epochs.

use_gpu::Bool = true

verbose::Bool = true

Arguments common to train(θ̂, P, simulator) and train(θ̂, θ_train, θ_val, simulator):

m: sample sizes (either an Integer or a collection of Integers).

epochs_per_Z_refresh::Integer = 1: how often to refresh the training data.

simulate_just_in_time::Bool = false: flag indicating whether we should simulate just-in-time, in the sense that only a batchsize number of parameter vectors and corresponding data are in memory at a given time.

Arguments unique to train(θ̂, P, simulator):

K::Integer = 10000: number of parameter vectors in the training set; the size of the validation set is K ÷ 5.

ξ = nothing: an arbitrary collection of objects that are fixed (e.g., distance matrices); if ξ is provided, the parameter sampler is called as sampler(K, ξ).

epochs_per_θ_refresh::Integer = 1: how often to refresh the training parameters. Must be a multiple of epochs_per_Z_refresh.

Examples

using NeuralEstimators
using Flux
using Distributions
import NeuralEstimators: simulate

# data simulator
m = 15
function simulate(θ_set, m)
	[Float32.(rand(Normal(θ[1], θ[2]), 1, m)) for θ ∈ eachcol(θ_set)]
end

# parameter sampler
K = 10000
Ω = (μ = Normal(0, 1), σ = Uniform(0.1, 1)) # prior
struct sampler
	μ
	σ
end
function (s::sampler)(K)
	μ = rand(s.μ, K)
	σ = rand(s.σ, K)
	θ = hcat(μ, σ)'
	return θ
end
smplr = sampler(Ω.μ, Ω.σ)

# architecture
p = length(Ω)   # number of parameters in the statistical model
w = 32          # width of each layer
ψ = Chain(Dense(1, w, relu), Dense(w, w, relu))
ϕ = Chain(Dense(w, w, relu), Dense(w, p))
θ̂ = DeepSet(ψ, ϕ)

# training: full simulation on-the-fly
θ̂ = train(θ̂, smplr, simulate, m = m, K = K, epochs = 5)
θ̂ = train(θ̂, smplr, simulate, m = m, K = K, epochs = 5)
θ̂ = train(θ̂, smplr, simulate, m = m, K = K, epochs = 10, epochs_per_θ_refresh = 4, epochs_per_Z_refresh = 2)

# training: simulation on-the-fly but with fixed parameters
θ_train = smplr(K)
θ_val   = smplr(K ÷ 5)
θ̂ = train(θ̂, θ_train, θ_val, simulate, m = m, epochs = 5)
θ̂ = train(θ̂, θ_train, θ_val, simulate, m = m, epochs = 5, epochs_per_Z_refresh = 2)

# training: fixed parameters and fixed data
Z_train = simulate(θ_train, m)
Z_val   = simulate(θ_val, m)
θ̂ = train(θ̂, θ_train, θ_val, Z_train, Z_val, epochs = 5)

NeuralEstimators.trainx — Function

trainx(θ̂, P, simulator, M; )
trainx(θ̂, θ_train, θ_val, simulator, M; )
trainx(θ̂, θ_train, θ_val, Z_train::T, Z_val::T, M; )
trainx(θ̂, θ_train, θ_val, Z_train::V, Z_val::V; ) where {V <: AbstractVector{AbstractVector{Any}}}

A wrapper around train to construct neural estimators for different sample sizes.

The collection M specifies the desired sample sizes. Each estimator is pre-trained with the estimator for the previous sample size. For example, if M = [m₁, m₂], the estimator for sample size m₂ is pre-trained with the estimator for sample size m₁.

The method for Z_train::T and Z_val::T subsets the data using subsetdata(Z, 1:mᵢ) for each mᵢ ∈ M. The method for Z_train::V and Z_val::V trains an estimator for each element of Z_train and Z_val; hence, it does not need to invoke subsetdata, which can be slow or difficult to define in some cases (e.g., for graphical data).

The keyword arguments inherit from train, and certain keyword arguments can be given as vectors. For example, if we are training two estimators, we can use a different number of epochs by providing epochs = [e₁, e₂]. Other arguments that allow vectors are batchsize, stopping_epochs, and optimiser.

Assessing a neural estimator

NeuralEstimators.assess — Function

assess(estimators, θ, Z; <keyword args>)

Using a collection of estimators, compute estimates from data Z simulated based on true parameter vectors stored in θ.

If Z contains more data sets than parameter vectors, the parameter matrix will be recycled by horizontal concatenation.

The output is of type Assessment; see ?Assessment for details.

Keyword arguments

estimator_names::Vector{String}: names of the estimators (sensible defaults provided).

parameter_names::Vector{String}: names of the parameters (sensible defaults provided). If ξ is provided with a field parameter_names, those names will be used.

ξ = nothing: an arbitrary collection of objects that are fixed (e.g., distance matrices).

use_ξ = false: a Bool or a collection of Bool objects with length equal to the number of estimators. Specifies whether or not the estimator uses ξ: if it does, the estimator will be applied as estimator(Z, ξ). This argument is useful when multiple estimators are provided, only some of which need ξ; hence, if only one estimator is provided and ξ is not nothing, use_ξ is automatically set to true.

use_gpu = true: a Bool or a collection of Bool objects with length equal to the number of estimators.

verbose::Bool = true

Examples

using NeuralEstimators
using Flux

n = 10 # number of observations in each realisation
p = 4  # number of parameters in the statistical model

# Construct the neural estimator
w = 32 # width of each layer
ψ = Chain(Dense(n, w, relu), Dense(w, w, relu));
ϕ = Chain(Dense(w, w, relu), Dense(w, p));
θ̂ = DeepSet(ψ, ϕ)

# Generate testing parameters
K = 100
θ = rand(p, K)

# Data for a single sample size
m = 30
Z = [rand(n, m) for _ ∈ 1:K];
assessment = assess([θ̂], θ, Z);
risk(assessment)

# Multiple data sets for each parameter vector
J = 5
Z = repeat(Z, J);
assessment = assess([θ̂], θ, Z);
risk(assessment)

# With set-level information
qₓ = 2
ϕ  = Chain(Dense(w + qₓ, w, relu), Dense(w, p));
θ̂ = DeepSet(ψ, ϕ)
x = [rand(qₓ) for _ ∈ eachindex(Z)]
assessment = assess([θ̂], θ, (Z, x));
risk(assessment)

NeuralEstimators.Assessment — Type

Assessment(df::DataFrame, runtime::DataFrame)

A type for storing the output of assess(). The field runtime contains the total time taken for each estimator. The field df is a long-form DataFrame with columns:

estimator: the name of the estimator

parameter: the name of the parameter

truth: the true value of the parameter

estimate: the estimated value of the parameter

m: the sample size (number of iid replicates)

k: the index of the parameter vector in the test set

j: the index of the data set

Multiple Assessment objects can be combined with merge().

NeuralEstimators.risk — Function

risk(assessment::Assessment; loss = (x, y) -> abs(x - y), average_over_parameters = true)

Computes a Monte Carlo approximation of the Bayes risk,

\[r_{\Omega}(\hat{\boldsymbol{\theta}}(\cdot)) \approx \frac{1}{K} \sum_{\boldsymbol{\theta} \in \vartheta} \frac{1}{J} \sum_{\boldsymbol{Z} \in \mathcal{Z}_{\boldsymbol{\theta}}} L(\boldsymbol{\theta}, \hat{\boldsymbol{\theta}}(\boldsymbol{Z})).\]

where $\vartheta$ denotes a set of $K$ parameter vectors sampled from the prior $\Omega(\cdot)$ and, for each $\boldsymbol{\theta} \in \vartheta$, we have $J$ sets of $m$ mutually independent realisations from the model collected in $\mathcal{Z}_{\boldsymbol{\theta}}$.

Keyword arguments

loss = (x, y) -> abs(x - y): a binary operator defining the loss function (default absolute-error loss).

average_over_parameters::Bool = true: if true (default), the loss is averaged over all parameters; otherwise, the loss is averaged over each parameter separately.

average_over_sample_sizes::Bool = true: if true (default), the loss is averaged over all sample sizes $m$; otherwise, the loss is averaged over each sample size separately.

Bootstrapping

NeuralEstimators.bootstrap — Function

bootstrap(θ̂, parameters::P, Z) where P <: Union{AbstractMatrix, ParameterConfigurations}
bootstrap(θ̂, parameters::P, simulator, m::Integer; B = 400) where P <: Union{AbstractMatrix, ParameterConfigurations}
bootstrap(θ̂, Z; B = 400, blocks = nothing)

Generates B bootstrap estimates from an estimator θ̂.

Parametric bootstrapping is facilitated by passing a single parameter configuration, parameters, and corresponding simulated data, Z, whose length implicitly defines B. Alternatively, one may provide a simulator and the desired sample size, in which case the data will be simulated using simulator(parameters, m).

Non-parametric bootstrapping is facilitated by passing a single data set, Z. The argument blocks caters for block bootstrapping, and it should be a vector of integers specifying the block for each replicate. For example, with 5 replicates, the first two corresponding to block 1 and the remaining three corresponding to block 2, blocks should be [1, 1, 2, 2, 2]. The resampling algorithm aims to produce resampled data sets that are of a similar size to Z, but this can only be achieved exactly if all blocks are equal in length.

The keyword argument use_gpu is a flag determining whether to use the GPU, if it is available (default true).

The return type is a p × B matrix, where p is the number of parameters in the model.

NeuralEstimators.interval — Function

interval(θ̃, θ̂ = nothing; type::String, probs = [0.05, 0.95], parameter_names)

Compute a confidence interval using the p × B matrix of bootstrap samples, θ̃, where p is the number of parameters in the model.

If type = "quantile", the interval is constructed by simply taking the quantiles of θ̃, and if type = "reverse-quantile", the so-called reverse-quantile method is used. In both cases, the quantile levels are controlled by the argument probs.

The rows can be named with a vector of strings parameter_names.

The return type is a p × 2 matrix, whose first and second columns respectively contain the lower and upper bounds of the interval.

Examples

using NeuralEstimators
p = 3
B = 50
θ̃ = rand(p, B)
θ̂ = rand(p)
interval(θ̃)
interval(θ̃, θ̂, type = "basic")