Core functions

This page documents the functions that are central to the workflow of NeuralEstimators. Its organisation reflects the order in which these functions appear in a standard implementation; that is, from sampling parameters from the prior distribution, to uncertainty quantification of the final estimates via bootstrapping.

Sampling parameters

Parameters sampled from the prior distribution $\Omega(\cdot)$ are stored as a $p \times K$ matrix, where $p$ is the number of parameters in the model and $K$ is the number of parameter vectors sampled from the prior distribution.

It can sometimes be helpful to wrap the parameter matrix in a user-defined type that also stores expensive intermediate objects needed for data simulated (e.g., Cholesky factors). In this case, the user-defined type should be a subtype of the abstract type ParameterConfigurations, whose only requirement is a field θ that stores the matrix of parameters. See Storing expensive intermediate objects for data simulation for further discussion.

NeuralEstimators.ParameterConfigurations — Type

ParameterConfigurations

An abstract supertype for user-defined types that store parameters and any intermediate objects needed for data simulation with simulate.

The user-defined type must have a field θ that stores the $p$ × $K$ matrix of parameters, where $p$ is the number of parameters in the model and $K$ is the number of parameter vectors sampled from the prior distribution. There are no other restrictions.

See subsetparameters for the generic function for subsetting these objects.

Examples

struct P <: ParameterConfigurations
	θ
	# ...
end

Simulating data

NeuralEstimators facilitates neural estimation for arbitrary statistical models by having the user implicitly define the model via simulated data. The user may provide simulated data directly, or provide a function that simulates data from the model (by overloading the generic function simulate).

The data should be stored as a Vector{A}, where each element of the vector is associated with one parameter configuration, and where A depends on the representation of the neural estimator. For example, if the neural estimator is a DeepSet object, the data should be stored as a Vector{Array}, where each array may store independent replicates in its final dimension. Similarly, if the neural estimator is a GNNEstimator, the data should be stored as a Vector{GNNGraph}, where each graph may store independent replicates in sub-graphs.

NeuralEstimators.simulate — Function

Generic function that may be overloaded to implicitly define a statistical model. Specifically, the user should provide a method simulate(parameters, m) that returns m simulated replicates for each element in the given set of parameters.

NeuralEstimators.simulate — Method

simulate(parameters, m, J::Integer)

Simulates J sets of m independent replicates for each parameter vector in parameters by calling simulate(parameters, m) a total of J times.

Neural-estimator representations

Although the user is free to construct their neural estimator however they see fit, NeuralEstimators provides several useful representations described below.

NeuralEstimators.DeepSet — Type

DeepSet(ψ, ϕ, a)
DeepSet(ψ, ϕ; a::String = "mean")

A neural estimator in the DeepSet representation,

\[θ̂(𝐙) ≡ ϕ(𝐚(\{ψ(𝐙ᵢ) : i = 1, …, m\})),\]

where 𝐙 ≡ (𝐙₁', …, 𝐙ₘ')' are independent replicates from the model, ψ and ϕ are neural networks, and 𝐚 is a permutation-invariant aggregation function.

The function 𝐚 must aggregate over the last dimension of an array (i.e., the replicates dimension). It can be specified as a positional argument of type Function, or as a keyword argument of type String with permissible values "mean", "sum", and "logsumexp".

Examples

using NeuralEstimators
using Flux

n = 10 # number of observations in each realisation
p = 4  # number of parameters in the statistical model

# Construct the neural estimator
w = 32 # width of each layer
ψ = Chain(Dense(n, w, relu), Dense(w, w, relu));
ϕ = Chain(Dense(w, w, relu), Dense(w, p));
θ̂ = DeepSet(ψ, ϕ)

# Apply the estimator to a single set of 3 realisations:
Z = rand(n, 3);
θ̂(Z)

# Apply the estimator to two sets each containing 3 realisations:
Z = [rand(n, m) for m ∈ (3, 3)];
θ̂(Z)

# Apply the estimator to two sets containing 3 and 4 realisations, respectively:
Z = [rand(n, m) for m ∈ (3, 4)];
θ̂(Z)

NeuralEstimators.PiecewiseEstimator — Type

PiecewiseEstimator(estimators, m_breaks)

Creates a piecewise estimator from a collection of estimators, based on the collection of sample-size changepoints, m_breaks, which should contain one element fewer than the number of estimators.

Examples

# Suppose that we've trained two neural estimators. The first, θ̂₁, is trained
# for small sample sizes (e.g., m ≤ 30), and the second, `θ̂₂`, is trained for
# moderate-to-large sample sizes (e.g., m > 30). Then we construct a piecewise
# estimator with a sample-size changepoint of 30, which dispatches θ̂₁ if m ≤ 30
# and θ̂₂ if m > 30.

n = 2  # bivariate data
p = 3  # number of parameters in the model
w = 8  # width of each layer

ψ₁ = Chain(Dense(n, w, relu), Dense(w, w, relu));
ϕ₁ = Chain(Dense(w, w, relu), Dense(w, p));
θ̂₁ = DeepSet(ψ₁, ϕ₁)

ψ₂ = Chain(Dense(n, w, relu), Dense(w, w, relu), Dense(w, w, relu));
ϕ₂ = Chain(Dense(w, w, relu), Dense(w, w, relu), Dense(w, p));
θ̂₂ = DeepSet(ψ₂, ϕ₂)

θ̂ = PiecewiseEstimator([θ̂₁, θ̂₂], [30])
Z = [rand(n, 1, m) for m ∈ (10, 50)]
θ̂(Z)

NeuralEstimators.GNNEstimator — Type

GNNEstimator(propagation, globalpool, deepset)

A neural estimator based on a graph neural network (GNN). The propagation module transforms graphical input data into a set of hidden feature graphs; the globalpool module aggregates the feature graphs (graph-wise) into a single hidden feature vector; and the deepset module maps the hidden feature vectors onto the parameter space.

The data should be a GNNGraph or AbstractVector{GNNGraph}, where each graph is associated with a single parameter vector. The graphs may contain sub-graphs corresponding to independent replicates from the model.

Examples

using NeuralEstimators
using Flux
using Flux: batch
using GraphNeuralNetworks
using Statistics: mean

# Create some graphs
d = 1             # dimension of the response variable
n₁, n₂ = 11, 27   # number of nodes
e₁, e₂ = 30, 50   # number of edges
g₁ = rand_graph(n₁, e₁, ndata=rand(d, n₁))
g₂ = rand_graph(n₂, e₂, ndata=rand(d, n₂))
g  = batch([g₁, g₂])

# propagation module
w = 5; o = 7
propagation = GNNChain(GraphConv(d => w), GraphConv(w => w), GraphConv(w => o))

# global pooling module
meanpool = GlobalPool(mean)

# Deep Set module
w = 32
p = 3
ψ₂ = Chain(Dense(o, w, relu), Dense(w, w, relu), Dense(w, w, relu))
ϕ₂ = Chain(Dense(w, w, relu), Dense(w, p))
deepset = DeepSet(ψ₂, ϕ₂)

# GNN estimator
est = GNNEstimator(propagation, meanpool, deepset)

# Apply the estimator to a single graph, a single graph containing sub-graphs,
# and a vector of graphs:
θ̂ = est(g₁)
θ̂ = est(g)
θ̂ = est([g₁, g₂, g])

Training

NeuralEstimators.train — Function

Generic function for training a neural estimator.

The methods are designed to cater for different forms of "on-the-fly simulation" (see the online documentation). In all methods, the validation data are held fixed to reduce noise when evaluating the validation risk function (which is used to monitor the performance of the estimator during training).

Keyword arguments

Arguments common to all methods:

loss = mae: the loss function, which should return the average loss when applied to multiple replicates.

epochs::Integer = 100

batchsize::Integer = 32

optimiser = ADAM(1e-4)

savepath::String = "": path to save the trained θ̂ and other information; if savepath is an empty string (default), nothing is saved.

stopping_epochs::Integer = 5: cease training if the risk doesn't improve in stopping_epochs epochs.

use_gpu::Bool = true

verbose::Bool = true

Arguments common to train(θ̂, P) and train(θ̂, θ_train, θ_val):

m: sample sizes (either an Integer or a collection of Integers).

epochs_per_Z_refresh::Integer = 1: how often to refresh the training data.

simulate_just_in_time::Bool = false: should we simulate the data "just-in-time"? If true, the user must overload the generic function simulate with a method simulate(parameters, m).

Arguments unique to train(θ̂, P):

K::Integer = 10000: number of parameter vectors in the training set; the size of the validation set is K ÷ 5.

ξ = nothing: an arbitrary collection of objects that are fixed (e.g., distance matrices); if ξ is provided, the constructor P is called as P(K, ξ).

NeuralEstimators.train — Method

train(θ̂, P; <keyword args>)

Train the neural estimator θ̂ by providing a constructor, P, where P <: Union{AbstractMatrix, ParameterConfigurations}, to automatically sample training and validation parameter sets at each epoch.

NeuralEstimators.train — Method

train(θ̂, θ_train::P, θ_val::P; <keyword args>)

Train the neural estimator θ̂ by providing the training and validation parameter sets explicitly as θ_train and θ_val, both of which are held fixed during training.

NeuralEstimators.train — Method

train(θ̂, θ_train::P, θ_val::P, Z_train::T, Z_val::T; <keyword args>)

Train the neural estimator θ̂ by providing the training and validation parameter sets, θ_train and θ_val, and the training and validation data sets, Z_train and Z_val, all of which are held fixed during training.

If the elements of Z_train and Z_val are equally replicated, and the number of replicates for each element of Z_train is a multiple of the number of replicates for each element of Z_val, then the training data will then be recycled throughout training to imitate on-the-fly simulation. For example, if each element of Z_train consists of 50 replicates, and each element of Z_val consists of 10 replicates, the first epoch uses the first 10 replicates in Z_train, the second epoch uses the next 10 replicates, and so on, until epoch 6 again uses the first 10 replicates. Note that this requires the data to be subsetted throughout training with the function subsetdata.

NeuralEstimators.train — Method

train(θ̂, θ_train::P, θ_val::P, Z_train::T, Z_val::T, M; <keyword args>)
train(θ̂, θ_train::P, θ_val::P, Z_train::V, Z_val::V; <keyword args>) where {V <: Vector{Vector{T}}}

Train several neural estimators with the architecture θ̂ under different sample sizes.

The first method accepts sample sizes as a vector of integers, M. Each estimator is pre-trained with the estimator trained for the previous sample size. For example, if M = [m₁, m₂], with m₂ > m₁, the estimator for sample size m₂ is pre-trained with the estimator for sample size m₁.

The second method requires Z_train and Z_val to be a Vector{Vector{T}}, where T is arbitrary. In this method, a separate estimator is trained for each element in Z_train and Z_val and, importantly, it does not invoke subsetdata(), which can be slow for graphical data. Again, pre-training is used.

These methods wrap train(θ̂, θ_train, θ_val, Z_train, Z_val) and, hence, they inherit its keyword arguments. Further, certain keyword arguments can be given as vectors. For instance, if we are training two neural estimators, we can use a different number of epochs by providing epochs = [e₁, e₂]. Other arguments that allow vectors are batchsize, stopping_epochs, and optimiser.

Assessing a neural estimator

NeuralEstimators.assess — Function

assess(estimators, parameters, Z; <keyword args>)
assess(estimators, parameters; <keyword args>)

Using a collection of estimators, compute estimates from data simulated from a set of parameters.

Testing data can be automatically simulated by overloading simulate with a method simulate(parameters, m::Integer), and using the keyword argument m to specify the desired sample sizes to assess. Alternatively, one may provide testing data Z as an iterable collection, where each element contains the testing data for a given sample size. If there are more simulated data sets than unique parameter vectors, the data should be stored in an 'outer' fashion, so that the parameter vectors run faster than the replicated data.

Keyword arguments

Arguments common to both methods

estimator_names::Vector{String}: names of the estimators (sensible defaults provided).

parameter_names::Vector{String}: names of the parameters (sensible defaults provided).

ξ = nothing: an arbitrary collection of objects that are fixed (e.g., distance matrices).

use_ξ = false: a Bool or a collection of Bool objects with length equal to the number of estimators. Specifies whether or not the estimator uses ξ: if it does, the estimator will be applied as estimator(Z, ξ). This argument is useful when multiple estimators are provided, only some of which need ξ; hence, if only one estimator is provided and ξ is not nothing, use_ξ is automatically set to true.

use_gpu = true: a Bool or a collection of Bool objects with length equal to the number of estimators.

verbose::Bool = true

Arguments unique to assess(estimators, parameters)

m::Vector{Integer}: sample sizes to estimate from.

J::Integer = 1: the number of times to replicate each parameter in parameters.

Examples

using NeuralEstimators
using Flux

n = 10 # number of observations in each realisation
p = 4  # number of parameters in the statistical model

# Construct the neural estimator
w = 32 # width of each layer
ψ = Chain(Dense(n, w, relu), Dense(w, w, relu));
ϕ = Chain(Dense(w, w, relu), Dense(w, p));
θ̂ = DeepSet(ψ, ϕ)

# Generate fake parameters and corresponding data for a range of sample sizes:
K = 100        # number of parameter vectors in the test set
θ = rand(p, K)
Z = [[rand(n, m) for _ ∈ 1:K] for m ∈ (1, 10, 20)]

assessment = assess([θ̂], θ, Z)
risk(assessment)
risk(assessment, average_over_parameters = false)

NeuralEstimators.Assessment — Type

Assessment(θandθ̂::DataFrame, runtime::DataFrame)

An type for storing the output of assess(). It contains two fields. The field runtime contains the total time taken for each estimator for each sample size m. The field θandθ̂ is a long-form DataFrame containing the true parameters and corresponding estimates. Specifically, its columns are:

estimator: the name of the estimator

parameter: the name of the parameter

truth: the true value of the parameter

estimate: the estimated value of the parameter

m: the sample size

k: the index of the parameter vector in the test set

j: the index of the data set

Multiple Assessment objects can be combined with the function merge.

NeuralEstimators.risk — Function

risk(assessment::Assessment; loss = (x, y) -> abs(x - y), average_over_parameters = true)

Estimates the Bayes risk with respect to the loss function for each estimator, parameter, and sample size considered in assessment.

The argument loss should be a binary operator (default absolute-error loss).

If average_over_parameters = true (default), the risk is averaged over all parameters; otherwise, the risk is evaluated over each parameter separately.

NeuralEstimators.coverage — Function

coverage(θ̂, Z::V, θ, α; kwargs...) where  {V <: AbstractArray{A}} where A

For each data set contained in Z, compute a non-parametric bootstrap confidence interval with nominal coverage α, and determine if the true parameters, θ, are contained within this interval. The overall empirical coverage is then obtained by averaging the resulting 0-1 matrix over all data sets.

NeuralEstimators.plotrisk — Function

TODO

NeuralEstimators.plotdistribution — Function

TODO

Bootstrapping

NeuralEstimators.bootstrap — Function

bootstrap(θ̂, parameters::P, Z̃) where P <: Union{AbstractMatrix, ParameterConfigurations}
bootstrap(θ̂, parameters::P, m::Integer; B = 400) where P <: Union{AbstractMatrix, ParameterConfigurations}
bootstrap(θ̂, Z; B = 400)
bootstrap(θ̂, Z, blocks::Vector{Integer}; B = 400)

Generates B bootstrap estimates from an estimator θ̂.

Parametric bootstrapping is facilitated by passing a single parameter configuration, parameters, and corresponding simulated data, Z̃, whose length implicitly defines B. Alternatively, if the user has defined a method simulate(parameters, m), one may simply pass the desired sample size m for the simulated data sets.

Non-parametric bootstrapping is facilitated by passing a single data set, Z. The argument blocks caters for block bootstrapping, and it should be an integer vector specifying the block for each replicate. For example, with 5 replicates, the first two corresponding to block 1 and the remaining three corresponding to block 2, blocks should be [1, 1, 2, 2, 2]. The resampling algorithm aims to produce resampled data sets that are of a similar size to Z, but this can only be achieved exactly if all blocks are equal in length.

The keyword argument use_gpu is a flag determining whether to use the GPU, if it is available (default true).

The return type is a p × B matrix, where p is the number of parameters in the model.

NeuralEstimators.confidenceinterval — Function

confidenceinterval(θ̃; probs = [0.05, 0.95], parameter_names)

Compute a confidence interval using the quantiles of the p × B matrix of bootstrap samples, θ̃, where p is the number of parameters in the model.

The quantile levels are controlled with the argument probs. The rows can be named with parameter_names (sensible defaults provided), which shold be a vector.

The return type is a p × 2 matrix, whose first and second columns respectively contain the lower and upper bounds of the confidence interval.