Loss functions

In addition to the standard loss functions provided by Flux (e.g., mae, mse, etc.), NeuralEstimators provides the following loss functions.

NeuralEstimators.kpowerloss — Function

kpowerloss(θ̂, y, k; agg = mean, safeorigin = true, ϵ = 0.1)

For k ∈ (0, ∞), the k-th power absolute-distance loss,

\[L(θ̂, θ) = |θ̂ - θ|ᵏ,\]

contains the squared-error, absolute-error, and 0-1 loss functions as special cases (the latter obtained in the limit as k → 0).

It is Lipschitz continuous iff k = 1, convex iff k ≥ 1, and strictly convex iff k > 1. It is quasiconvex for all k > 0.

If safeorigin = true, the loss function is modified to avoid pathologies around the origin, so that the resulting loss function behaves similarly to the absolute-error loss in the ϵ-interval surrounding the origin.

NeuralEstimators.quantileloss — Function

quantileloss(θ̂, θ, q; agg = mean)
quantileloss(θ̂, θ, q::V; agg = mean) where {T, V <: AbstractVector{T}}

The asymmetric loss function whose minimiser is the qth posterior quantile; namely,

\[L(θ̂, θ, q) = (θ̂ - θ)(𝕀(θ̂ - θ > 0) - q),\]

where q ∈ (0, 1) and 𝕀(⋅) is the indicator function.

The method that takes q as a vector is useful for jointly approximating several quantiles of the posterior distribution. In this case, the number of rows in θ̂ is assumed to be pr, where p is the number of parameters: then, q should be an r-vector.

For further discussion on this loss function, see Equation (7) of Cressie, N. (2022), "Decisions, decisions, decisions in an uncertain environment", arXiv:2209.13157.

Examples

p = 1
K = 10
θ = rand(p, K)
θ̂ = rand(p, K)
quantileloss(θ̂, θ, 0.1)

θ̂ = rand(3p, K)
quantileloss(θ̂, θ, [0.1, 0.5, 0.9])

p = 2
θ = rand(p, K)
θ̂ = rand(p, K)
quantileloss(θ̂, θ, 0.1)

θ̂ = rand(3p, K)
quantileloss(θ̂, θ, [0.1, 0.5, 0.9])

NeuralEstimators.intervalscore — Function

intervalscore(l, u, θ, α; agg = mean)
intervalscore(θ̂, θ, α; agg = mean)

Given a 100×(1-α)% confidence interval [l, u] with true value θ, the interval score is defined by

\[S(l, u, θ; α) = (u - l) + 2α⁻¹(l - θ)𝕀(θ < l) + 2α⁻¹(θ - u)𝕀(θ > u),\]

where α ∈ (0, 1) and 𝕀(⋅) is the indicator function.

The method that takes a single value θ̂ assumes that θ̂ is a matrix with 2p rows, where p is the number of parameters in the statistical model. Then, the first and second set of p rows will be used as l and u, respectively.

For further discussion, see Section 6 of Gneiting, T. and Raftery, A. E. (2007), "Strictly proper scoring rules, prediction, and estimation", Journal of the American statistical Association, 102, 359–378.