Model-specific functions

Data simulators

The philosophy of NeuralEstimators is to cater for arbitrary statistical models by having the user define their statistical model implicitly through simulated data. However, the following functions have been included as they may be helpful to others, and their source code provide an example for how a user could formulate code for their own model. If you've developed similar functions that you think may be helpful to others, please get in touch or make a pull request.

NeuralEstimators.simulategaussianprocess — Function

simulategaussianprocess(L::Matrix, m = 1)
simulategaussianprocess(grf::GaussianRandomField, m = 1)

Simulates m independent and identically distributed (i.i.d.) realisations from a mean-zero Gaussian process.

Accepts either the lower Cholesky factor L associated with a Gaussian process or a GaussianRandomField object grf.

Examples

using NeuralEstimators

n  = 500
S  = rand(n, 2)
ρ  = 0.6
ν  = 1.0

# Passing GaussianRandomField object:
using GaussianRandomFields
cov = CovarianceFunction(2, Matern(ρ, ν))
grf = GaussianRandomField(cov, Cholesky(), S)
simulategaussianprocess(grf)

# Passing Cholesky factors directly as matrices:
L = grf.data
simulategaussianprocess(L)

# Circulant embedding, which is fast but can on only be used on grids:
pts = 1.0:50.0
grf = GaussianRandomField(cov, CirculantEmbedding(), pts, pts, minpadding = 100)
simulategaussianprocess(grf)

NeuralEstimators.simulateschlather — Function

simulateschlather(L::Matrix, m = 1)
simulateschlather(grf::GaussianRandomField, m = 1)

Simulates m independent and identically distributed (i.i.d.) realisations from Schlather's max-stable model using the algorithm for approximate simulation given by Schlather (2002), "Models for stationary max-stable random fields", Extremes, 5:33-44.

Accepts either the lower Cholesky factor L associated with a Gaussian process or a GaussianRandomField object grf.

Keyword arguments

C = 3.5: a tuning parameter that controls the accuracy of the algorithm: small C favours computational efficiency, while large C favours accuracy. Schlather (2002) recommends the use of C = 3.

Gumbel = true: flag indicating whether the data should be log-transformed from the unit Fréchet scale to the Gumbel scale.

Examples

using NeuralEstimators

n  = 500
S  = rand(n, 2)
ρ  = 0.6
ν  = 1.0

# Passing GaussianRandomField object:
using GaussianRandomFields
cov = CovarianceFunction(2, Matern(ρ, ν))
grf = GaussianRandomField(cov, Cholesky(), S)
simulateschlather(grf)

# Passing Cholesky factors directly as matrices:
L = grf.data
simulateschlather(L)

# Circulant embedding, which is fast but can on only be used on grids:
pts = 1.0:50.0
grf = GaussianRandomField(cov, CirculantEmbedding(), pts, pts, minpadding = 100)
simulateschlather(grf)

Low-level functions

These low-level functions may be of use for various models.

NeuralEstimators.matern — Function

matern(h, ρ, ν, σ² = 1)

For two points separated by h units, compute the Matérn covariance function, with range parameter ρ, smoothness parameter ν, and marginal variance parameter σ².

We use the parametrisation $C(\|\mathbf{h}\|) = \sigma^2 \frac{2^{1 - \nu}}{\Gamma(\nu)} \left(\frac{\|\mathbf{h}\|}{\rho}\right)^\nu K_\nu \left(\frac{\|\mathbf{h}\|}{\rho}\right)$, where $\Gamma(\cdot)$ is the gamma function, and $K_\nu(\cdot)$ is the modified Bessel function of the second kind of order $\nu$.

NeuralEstimators.maternchols — Function

maternchols(D, ρ, ν, σ² = 1)

Given a distance matrix D, constructs the Cholesky factor of the covariance matrix under the Matérn covariance function with range parameter ρ, smoothness parameter ν, and marginal variance σ².

Providing vectors of parameters will yield a three-dimensional array of Cholesky factors (note that the vectors must of the same length, but a mix of vectors and scalars is allowed). A vector of distance matrices D may also be provided.

Examples

using NeuralEstimators
using LinearAlgebra: norm
n  = 10
S  = rand(n, 2)
D  = [norm(sᵢ - sⱼ) for sᵢ ∈ eachrow(S), sⱼ ∈ eachrow(S)]
ρ  = [0.6, 0.5]
ν  = [0.7, 1.2]
σ² = [0.2, 0.4]
maternchols(D, ρ, ν)
maternchols(D, ρ, ν, σ²)

S̃  = rand(n, 2)
D̃  = [norm(sᵢ - sⱼ) for sᵢ ∈ eachrow(S̃), sⱼ ∈ eachrow(S̃)]
maternchols([D, D̃], ρ, ν, σ²)

Density functions

Density functions are not needed in the workflow of NeuralEstimators. However, as part of a series of comparison studies between neural estimators and likelihood-based estimators given in the manuscript, we have developed the following density functions, and we include them in NeuralEstimators to cater for the possibility that they may be of use in future comparison studies.

NeuralEstimators.gaussiandensity — Function

gaussiandensity(y::V, L; logdensity = true) where {V <: AbstractVector{T}} where T
gaussiandensity(y::A, Σ; logdensity = true) where {A <: AbstractArray{T, N}} where {T, N}

Efficiently computes the density function for y ~ 𝑁(0, Σ), with L the lower Cholesky factor of the covariance matrix Σ.

The method gaussiandensity(y::A, Σ) assumes that the last dimension of y corresponds to the independent-replicates dimension, and it exploits the fact that we need to compute the Cholesky factor L for these independent replicates once only.

NeuralEstimators.schlatherbivariatedensity — Function

schlatherbivariatedensity(z₁, z₂, ψ; logdensity = true)

The bivariate density function for Schlather's max-stable model, as given in Huser (2013, pg. 231–232).

Huser, R. (2013). Statistical Modeling and Inference for Spatio-Temporal Ex- tremes. PhD thesis, Swiss Federal Institute of Technology, Lausanne, Switzerland.