Miscellaneous

NeuralEstimators.NeuralEM

NeuralEstimators.adjacencymatrix

NeuralEstimators.containertype

NeuralEstimators.encodedata

NeuralEstimators.estimateinbatches

NeuralEstimators.expandgrid

NeuralEstimators.initialise_estimator

NeuralEstimators.loadbestweights

NeuralEstimators.numberreplicates

NeuralEstimators.removedata

NeuralEstimators.rowwisenorm

NeuralEstimators.stackarrays

NeuralEstimators.subsetdata

NeuralEstimators.subsetparameters

NeuralEstimators.vectotril

Core

These functions can appear during the core workflow, and may need to be overloaded in some applications.

NeuralEstimators.numberreplicates — Function

numberofreplicates(Z)

Generic function that returns the number of replicates in a given object. Default implementations are provided for commonly used data formats, namely, data stored as an Array or as a GNNGraph.

NeuralEstimators.subsetdata — Function

Generic function for subsetting replicates from a data set. Default methods are:

subsetdata(Z::A, m) where {A <: AbstractArray{T, N}} where {T, N}
subsetdata(Z::G, m) where {G <: AbstractGraph}

Note that subsetdata is slow for graphical data, and one should consider using a method of train that does not require the data to be subsetted when working with graphical data: use numberreplicates to check that the training and validation data sets are equally replicated, which prevents the invocation of subsetdata. Note also that subsetdata only applies to vectors of batched graphs.

If the user is working with data that is not covered by the default methods, simply overload subsetdata with the appropriate type for Z.

Examples

using NeuralEstimators
using GraphNeuralNetworks
using Flux: batch

n = 5  # number of observations in each realisation
m = 6  # number of replicates for each parameter vector
d = 1  # dimension of the response variable
K = 2  # number of parameter vectors

# Array data
Z = [rand(n, d, m) for k ∈ 1:K]
subsetdata(Z, 1:3) # extract first 3 replicates for each parameter vector

# Graphical data
e = 8 # number of edges
Z = [batch([rand_graph(n, e, ndata = rand(d, n)) for _ ∈ 1:m]) for k ∈ 1:K]
subsetdata(Z, 1:3) # extract first 3 replicates for each parameter vector

NeuralEstimators.subsetparameters — Function

subsetparameters(parameters::M, indices) where {M <: AbstractMatrix}
subsetparameters(parameters::P, indices) where {P <: ParameterConfigurations}

Subset parameters using a collection of indices.

Arrays in parameters::P with last dimension equal in size to the number of parameter configurations, K, are also subsetted (over their last dimension) using indices. All other fields are left unchanged. To modify this default behaviour, overload subsetparameters.

Utility functions

NeuralEstimators.adjacencymatrix — Function

adjacencymatrix(M::Matrix, k::Integer)
adjacencymatrix(M::Matrix, r::Float)
adjacencymatrix(M::Matrix, r::Float, k::Integer)

Computes a spatially weighted adjacency matrix from M based on either the k nearest neighbours of each location, or a fixed spatial radius of r units; if both r and k are provided, randomly selects k neighbours within a radius of r units.

If M is a square matrix, is it treated as a distance matrix; otherwise, it should be an n x d matrix, where n is the number of spatial sample locations and d is the spatial dimension (typically d = 2). In the latter case, the distance metric is taken to be the Euclidean distance.

All methods accept the keyword argument self_loops (default false); set to true nodes are considered to self connected, so that the diagonal of the adjacency matrix is non-zero.

See also the package NearestNeighbors.jl.

Examples

using NeuralEstimators
using Distances

n = 100
d = 2
S = rand(n, d)
k = 5
r = 0.3

# Memory efficient constructors (avoids constructing the full distance matrix D)
adjacencymatrix(S, k)
adjacencymatrix(S, r)
adjacencymatrix(S, r, k)

# Construct from full distance matrix D
D = pairwise(Euclidean(), S, S, dims = 1)
adjacencymatrix(D, k)
adjacencymatrix(D, r)
adjacencymatrix(D, r, k)

NeuralEstimators.containertype — Function

containertype(A::Type)
containertype(::Type{A}) where A <: SubArray
containertype(a::A) where A

Returns the container type of its argument.

If given a SubArray, returns the container type of the parent array.

Examples

a = rand(3, 4)
containertype(a)
containertype(typeof(a))
[containertype(x) for x ∈ eachcol(a)]

NeuralEstimators.encodedata — Function

encodedata(Z::A; fixed_constant::T = zero(T)) where {A <: AbstractArray{Union{Missing, T}, N}} where T, N

For data Z with missing entries, returns an augmented data set (U, W) where W encodes the missingness pattern as an indicator vector and U is the original data Z with missing entries replaced by a fixed_constant.

The indicator vector W is stored in the second-to-last dimension of Z, which should be a singleton. If the second-to-last dimension is not singleton, then two singleton dimensions will be added to the array, and W will be stored in the new second-to-last dimension.

Examples

using NeuralEstimators

# Generate some missing data
Z = rand(16, 16, 1, 1)
Z = removedata(Z, 0.25)	 # remove 25% of the data

# Encode the data
UW = encodedata(Z)

NeuralEstimators.estimateinbatches — Function

estimateinbatches(θ̂, z; batchsize::Integer = 32, use_gpu::Bool = true)

Apply the estimator θ̂ on minibatches of z of size batchsize, to avoid memory issues that can occur when z is very large.

Minibatching will only be done if there are multiple data sets in z; this will be inferred by z being a vector, or a tuple whose first element is a vector.

NeuralEstimators.expandgrid — Function

expandgrid(xs, ys)

Same as expand.grid() in R, but currently caters for two dimensions only.

NeuralEstimators.initialise_estimator — Function

initialise_estimator(p::Integer; ...)

Initialise a neural estimator for a statistical model with p unknown parameters.

The estimator is couched in the DeepSets framework (see DeepSet) so that it can be applied to data sets containing an arbitrary number of independent replicates (including the special case of a single replicate).

Note also that the user is free to initialise their neural estimator however they see fit using arbitrary Flux code; see here for Flux's API reference.

Keyword arguments

architecture::String: for unstructured multivariate data, one may use a densely-connected neural network ("DNN"); for data collected over a grid, a convolutional neural network ("CNN"); and for graphical or irregular spatial data, a graphical neural network ("GNN").

d::Integer = 1: for unstructured multivariate data (i.e., when architecture = "DNN"), the dimension of the data (e.g., d = 3 for trivariate data); otherwise, if architecture ∈ ["CNN", "GNN"], the argument d controls the number of input channels (e.g., d = 1 for univariate spatial processes).

estimator_type::String = "point": the type of estimator; either "point" or "interval".

depth = 3: the number of hidden layers; either a single integer or an integer vector of length two specifying the depth of the inner (summary) and outer (inference) network of the DeepSets framework.

width = 32: a single integer or an integer vector of length sum(depth) specifying the width (or number of convolutional filters/channels) in each hidden layer.

activation::Function = relu: the (non-linear) activation function of each hidden layer.

activation_output::Function = identity: the activation function of the output layer.

kernel_size = nothing: (applicable only to CNNs) a vector of length depth[1] containing integer tuples of length D, where D is the dimension of the convolution (e.g., D = 2 for two-dimensional convolution).

weight_by_distance::Bool = false: (applicable only to GNNs) flag indicating whether the estimator will weight by spatial distance; if true, a WeightedGraphConv layer is used in the propagation module; otherwise, a regular GraphConv layer is used.

Examples

## DNN, GNN, 1D CNN, and 2D CNN for a statistical model with two parameters:
p = 2
initialise_estimator(p, architecture = "DNN")
initialise_estimator(p, architecture = "GNN")
initialise_estimator(p, architecture = "CNN", kernel_size = [10, 5, 3])
initialise_estimator(p, architecture = "CNN", kernel_size = [(10, 10), (5, 5), (3, 3)])

NeuralEstimators.loadbestweights — Function

loadbestweights(path::String)

Returns the weights of the neural network saved as 'best_network.bson' in the given path.

NeuralEstimators.removedata — Function

removedata(Z::Array, Iᵤ::Vector{Integer})
removedata(Z::Array, p::Union{Float, Vector{Float}}; prevent_complete_missing = true)
removedata(Z::Array, n::Integer; fixed_pattern = false, contiguous_pattern = false, variable_proportion = false)

Replaces elements of Z with missing.

The simplest method accepts are of integers Iᵤ that give the specific indices of the data to be removed.

Alterntivaly, there are two methods available to generate data that are missing completely at random (MCAR).

First, a vector p may be given that specifies the proportion of missingness for each element in the response vector. Hence, p should have length equal to the dimension of the response vector. If a single proportion is given, it will be replicated accordingly. If prevent_complete_missing = true, no replicates will contain 100% missingness (note that this can slightly alter the effective values of p).

Second, if an integer n is provided, all replicates will contain n observations after the data are removed. If fixed_pattern = true, the missingness pattern is fixed for all replicates. If contiguous_pattern = true, the data will be removed in a contiguous block. If variable_proportion = true, the proportion of missingness will vary across replicates, with each replicate containing between 1 and n observations after data removal, sampled uniformly (note that variable_proportion overrides fixed_pattern).

The return type is Array{Union{T, Missing}}.

Examples

d = 5           # dimension of each replicate
m = 2000        # number of replicates
Z = rand(d, m)  # simulated data

# Passing a desired proportion of missingness
p = rand(d)
removedata(Z, p)

# Passing a desired final sample size
n = 3  # number of observed elements of each replicate: must have n <= d
removedata(Z, n)

NeuralEstimators.rowwisenorm — Function

rowwisenorm(A)

Computes the row-wise norm of a matrix A.

NeuralEstimators.stackarrays — Function

stackarrays(v::V; merge = true) where {V <: AbstractVector{A}} where {A <: AbstractArray{T, N}} where {T, N}

Stack a vector of arrays v along the last dimension of each array, optionally merging the final dimension of the stacked array.

The arrays must be of the same size for the first N-1 dimensions. However, if merge = true, the size of the final dimension can vary.

Examples

# Vector containing arrays of the same size:
Z = [rand(2, 3, m) for m ∈ (1, 1)];
stackarrays(Z)
stackarrays(Z, merge = false)

# Vector containing arrays with differing final dimension size:
Z = [rand(2, 3, m) for m ∈ (1, 2)];
stackarrays(Z)

NeuralEstimators.vectotril — Function

vectotril(v; strict = false)
vectotriu(v; strict = false)

Converts a vector v of length $d(d+1)÷2$ (a triangular number) into a $d × d$ lower or upper triangular matrix.

If strict = true, the matrix will be strictly lower or upper triangular, that is, a $(d+1) × (d+1)$ triangular matrix with zero diagonal.

Note that the triangular matrix is constructed on the CPU, but the returned matrix will be a GPU array if v is a GPU array. Note also that the return type is not of type Triangular matrix (i.e., the zeros are materialised) since Traingular matrices are not always compatible with other GPU operations.

Examples

using NeuralEstimators

d = 4
n = d*(d+1)÷2
v = collect(range(1, n))
vectotril(v)
vectotriu(v)
vectotril(v; strict = true)
vectotriu(v; strict = true)

Other

NeuralEstimators.NeuralEM — Type

NeuralEM(simulateconditional::Function, neuralMAP::NeuralPointEstimator, θ₀ = nothing)
NeuralEM(neuralem::NeuralEM, θ₀)

A type that implements the neural expectation-maximisation (EM) algorithm using a function for conditional simulation over the missing values (simulateconditional), a neural approximation of the MAP estimator (neuralMAP), and a vector of starting parameter values (θ₀).

Fields of NeuralEM objects

The function simulateconditional should be of the form,

simulateconditional(Z::A, θ, ξ = nothing; nsims::Integer = 1) where {A <: AbstractArray{Union{Missing, T}}} where T

The data Z should be returned in whatever form is amenable to the architecture of the neural MAP estimator. For instance, if the data are gridded and the neural MAP estimator is based on a CNN architecture, then Z should be returned as a four-dimensional array.

The neuralMAP estimator should be a neural point estimator trained to approximate the joint posterior mode.

The starting value θ₀ should be a vector, and it can be provided either during construction of the NeuralEM object, or when applying the NeuralEM object to data (see below). The starting values given in a function call take precedence over those stored in the object.

Methods

Once constructed, obects of type NeuralEM can be applied to data via the method,

(neuralem::NeuralEM)(
	Z::A, θ₀ = nothing;
	niterations::Integer = 50,
	nsims::Integer = 1,
	ξ = nothing,
	ϵ = 0.01,
	return_iterates::Bool = false,
	use_gpu::Bool = true,
	verbose::Bool = false
)  where {A <: AbstractArray{Union{Missing, T}, N}} where {T, N}

The keyword arguments are:

Z: the complete-data vector containing the observed data Z₁ and Missing values in the missing component Z₂. The last dimension contains the replicates (if any); the other dimensions store the response variable.

θ₀: starting parameter values.

niterations: the maximum number of iterations to apply the algorithm for.

nsims: the number of conditional replicates used to approximate the conditional expectation (should align with the number of replicates that was used during training of the neural MAP estimator).

ξ: model information needed for conditional simulation (e.g., distance matrices).

ϵ: tolerance used to assess convergence.

return_iterates: if true, the estimate at each iteration of the algorithm is returned; otherwise, only the final estimate is returned.

The algorithm is stopped after niterations iterations or if the relative change in parameter values from successive iterations is sufficiently small, specifically, if

\[max_k (|θ_k^{(l+1)} - θ_k^{(l)}| / |θ_k^{(l)}|) < ϵ.\]

The following wrapper can be used when one has multiple data sets:

(neuralem::NeuralEM)(Z::V, θ₀::Union{Vector, Matrix, Nothing} = nothing; args...) where {V <: AbstractVector{A}} where {A <: AbstractArray{Union{Missing, T}, N}} where {T, N}

Here, the starting values θ₀ will be repeated if given as a vector.

Examples

# Please see the example given in the main documentation.