Miscellaneous

Core

These functions can appear during the core workflow, and may need to be overloaded in some applications.

NeuralEstimators.subsetparameters — Function

subsetparameters(parameters::M, indices) where {M <: AbstractMatrix}
subsetparameters(parameters::P, indices) where {P <: ParameterConfigurations}

Subset parameters using a collection of indices.

Arrays in parameters::P with last dimension equal in size to the number of parameter configurations, K, are also subsetted (over their last dimension) using indices. All other fields are left unchanged. To modify this default behaviour, overload subsetparameters.

NeuralEstimators.numberreplicates — Function

numberofreplicates(Z)

Generic function that returns the number of replicates in a given object. Default implementations are provided for commonly used data formats, namely, data stored as an Array or as a GNNGraph.

NeuralEstimators.subsetdata — Function

Generic function for subsetting replicates from a data set. Default methods are:

subsetdata(Z::A, m) where {A <: AbstractArray{T, N}} where {T, N}
subsetdata(Z::G, m) where {G <: AbstractGraph}

Note that subsetdata is slow for graphical data, and one should consider using a method of train that does not require the data to be subsetted. Use numberreplicates to check that the training and validation data sets are equally replicated, which prevents the invocation of subsetdata. Note also that subsetdata only applies to vectors of batched graphs.

If the user is working with data that is not covered by the default methods, simply overload subsetdata with the appropriate type for Z.

Examples

using NeuralEstimators
using GraphNeuralNetworks
using Flux: batch

n = 5  # number of observations in each realisation
m = 6  # number of replicates for each parameter vector
d = 1  # dimension of the response variable
K = 2  # number of parameter vectors

# Array data
Z = [rand(n, d, m) for k ∈ 1:K]
subsetdata(Z, 1:3) # extract first 3 replicates for each parameter vector

# Graphical data
e = 8 # number of edges
Z = [batch([rand_graph(n, e, ndata = rand(d, n)) for _ ∈ 1:m]) for k ∈ 1:K]
subsetdata(Z, 1:3) # extract first 3 replicates for each parameter vector

Architecture layers

These layers can be used at the end of an architecture to ensure that the neural estimator provides valid parameters.

NeuralEstimators.Compress — Type

Compress(a, b)

Uses the scaled logistic function to compress the output of a neural network to be between a and b.

The elements of a should be less than the corresponding element of b.

Examples

using NeuralEstimators
using Flux

p = 3
a = [0.1, 4, 2]
b = [0.9, 9, 3]
l = Compress(a, b)
K = 10
θ = rand(p, K)
l(θ)

n = 20
Z = rand(n, K)
θ̂ = Chain(Dense(n, 15), Dense(15, p), l)
θ̂(Z)

NeuralEstimators.CholeskyParameters — Type

CholeskyParameters(d)
CholeskyParametersConstrained(d, determinant = 1f0)

Layer for constructing the parameters of a Cholesky factor for a d-dimensional random vector.

This layer transforms an Matrix with d(d+1)÷2 rows (the number of non-zero elements in a Cholesky factor) into a Matrix of the same dimension, but with d rows constrained to be positive (corresponding to the diagonal elements of the Cholesky factor) and the remaining rows unconstrained.

The ordering of the transformed array aligns with Julia's column-major ordering, so that a Cholesky factor with d = 3,

\[\begin{bmatrix} L₁₁ & & \\ L₂₁ & L₂₂ & \\ L₃₁ & L₃₂ & L₃₃ \\ \end{bmatrix},\]

will follow the ordering $[L₁₁, L₂₁, L₃₁, L₂₂, L₃₂, L₃₃]'$. Since the diagonal elements must be positive, in this example rows 1, 4, and 6 of the transformed array will be constrained to be positive.

CholeskyParametersConstrained constrains the determinant of the Cholesky factor. Since the determinant of a triangular matrix is equal to the product of its diagonal elements, the determinant is constrained by setting the final diagonal element equal to determinant/$(Π Lᵢᵢ)$ where the product is over $i < d$.

Examples

using NeuralEstimators
using LinearAlgebra
using Test

d = 4
l = CholeskyParameters(d)
K = 50
p = d*(d+1)÷2
x = randn(p, K)
l(x)                                  # returns a matrix (used for Flux networks)
[vectotri(y) for y ∈ eachcol(l(x))]   # convert matrix to Cholesky factors

l = CholeskyParametersConstrained(d)
L = [vectotri(y) for y ∈ eachcol(l(x))]
@test all(det.(L) .≈ 1)

l = CholeskyParametersConstrained(d, 2f0)
L = [vectotri(y) for y ∈ eachcol(l(x))]
@test all(det.(L) .≈ 2)

NeuralEstimators.CovarianceMatrixParameters — Type

CovarianceMatrixParameters(d)
CovarianceMatrixParametersConstrained(d, determinant = 1f0)

Layer for constructing the parameters of a covariance matrix for a d-dimensional random vector.

Due to symmetry, there are d(d + 1)/2 free parameters in a covariance matrix, so this layer transforms a Matrix with d(d + 1)/2 rows into a Matrix of the same dimension. Internally, it uses a CholeskyParameters layer to construct a valid Cholesky factor, from which a positive-definite covariance matrix Σ can be computed.

The ordering of the transformed array aligns with Julia's column-major ordering, so that a covariance matrix with d = 3,

\[\begin{bmatrix} Σ₁₁ & Σ₁₂ & Σ₁₃ \\ Σ₂₁ & Σ₂₂ & Σ₂₃ \\ Σ₃₁ & Σ₃₂ & Σ₃₃ \\ \end{bmatrix},\]

will follow the ordering $[Σ₁₁, Σ₂₁, Σ₃₁, Σ₂₂, Σ₃₂, Σ₃₃]'$. Only the lower triangle of the matrix is returned because covariance matrices are symmetric.

CovarianceMatrixParametersConstrained constrains the determinant of the covariance matrix to determinant.

Examples

using NeuralEstimators
using LinearAlgebra
using Test

d = 4
l = CovarianceMatrixParameters(d)
K = 50
p = d*(d+1)÷2
x = randn(p, K)

l(x)
Σ = [Symmetric(vectotri(y), :L) for y ∈ eachcol(l(x))]
Σ = convert.(Matrix, Σ)
@test all(isposdef.(Σ))

l = CovarianceMatrixParametersConstrained(d)
Σ = [Symmetric(vectotri(y), :L) for y ∈ eachcol(l(x))]
Σ = convert.(Matrix, Σ)
@test all(det.(Σ) .≈ 1)

l = CovarianceMatrixParametersConstrained(d, 4f0)
Σ = [Symmetric(vectotri(y), :L) for y ∈ eachcol(l(x))]
Σ = convert.(Matrix, Σ)
@test all(det.(Σ) .≈ 4)

NeuralEstimators.SplitApply — Type

SplitApply(layers, indices)

Splits an array into multiple sub-arrays by subsetting the rows using the collection of indices, and then applies each layer in layers to the corresponding sub-array.

Specifically, for each i = 1, …, $n$, with $n$ the number of layers, SplitApply(x) performs layers[i](x[indices[i], :]), and then vertically concatenates the resulting transformed arrays.

Examples

using NeuralEstimators

d = 4
K = 50
p₁ = 2          # number of non-covariance matrix parameters
p₂ = d*(d+1)÷2  # number of covariance matrix parameters
p = p₁ + p₂

a = [0.1, 4]
b = [0.9, 9]
l₁ = Compress(a, b)
l₂ = CovarianceMatrixParameters(d)
l = SplitApply([l₁, l₂], [1:p₁, p₁+1:p])

θ = randn(p, K)
l(θ)

Utility functions

NeuralEstimators.vectotri — Function

vectotri(v, uplo = :L)

Converts a vector v of length $d(d+1)÷2$ into a $d$-dimensional triangular matrix.

Examples

d = 4
n = d*(d+1)÷2
v = range(1, n)
vectotri(v)
vectotri(v, :U)

NeuralEstimators.loadbestweights — Function

loadbestweights(path::String)

Given a path to a training run containing neural networks saved with names "network_epochx.bson" and an object saved as "loss_per_epoch.bson", returns the weights of the best network (measured by validation loss).

NeuralEstimators.stackarrays — Function

stackarrays(v::V; merge = true) where {V <: AbstractVector{A}} where {A <: AbstractArray{T, N}} where {T, N}

Stack a vector of arrays v along the last dimension of each array, optionally merging the final dimension of the stacked array.

The arrays must be of the same size for the first N-1 dimensions. However, if merge = true, the size of the final dimension can vary.

Examples

# Vector containing arrays of the same size:
Z = [rand(2, 3, m) for m ∈ (1, 1)];
stackarrays(Z)
stackarrays(Z, merge = false)

# Vector containing arrays with differing final dimension size:
Z = [rand(2, 3, m) for m ∈ (1, 2)];
stackarrays(Z)

NeuralEstimators.expandgrid — Function

expandgrid(xs, ys)

Same as expand.grid() in R, but currently caters for two dimensions only.