Miscellaneous

NeuralEstimators.EM

NeuralEstimators.adjacencymatrix

NeuralEstimators.containertype

NeuralEstimators.encodedata

NeuralEstimators.estimateinbatches

NeuralEstimators.expandgrid

NeuralEstimators.initialise_estimator

NeuralEstimators.loadbestweights

NeuralEstimators.numberreplicates

NeuralEstimators.removedata

NeuralEstimators.rowwisenorm

NeuralEstimators.stackarrays

NeuralEstimators.subsetdata

NeuralEstimators.subsetparameters

NeuralEstimators.vectotril

Core

These functions can appear during the core workflow, and may need to be overloaded in some applications.

NeuralEstimators.numberreplicates — Function

numberofreplicates(Z)

Generic function that returns the number of replicates in a given object. Default implementations are provided for commonly used data formats, namely, data stored as an Array or as a GNNGraph.

NeuralEstimators.subsetdata — Function

subsetdata(Z::V, m) where {V <: AbstractArray{A}} where {A <: Any}
subsetdata(Z::A, m) where {A <: AbstractArray{T, N}} where {T, N}
subsetdata(Z::G, m) where {G <: AbstractGraph}

Subsets m replicates from data Z.

Note that subsetdata is slow for graphical data, and one should consider using a method of train that does not require the data to be subsetted when working with graphical data: use numberreplicates to check that the training and validation data sets are equally replicated, which prevents the invocation of subsetdata. Note also that subsetdata only applies to vectors of batched graphs.

If the user is working with data that is not covered by the default methods, simply overload subsetdata with the appropriate type for Z.

Examples

using NeuralEstimators
using GraphNeuralNetworks
using Flux: batch

d = 1  # dimension of the response variable
n = 5  # number of observations in each realisation
m = 6  # number of replicates in each data set
K = 2  # number of data sets

# Array data
Z = [rand(n, d, m) for k ∈ 1:K]
subsetdata(Z, 1:3) # extract first 3 replicates from each data set

# Graphical data
e = 8 # number of edges
Z = [batch([rand_graph(n, e, ndata = rand(d, n)) for _ ∈ 1:m]) for k ∈ 1:K]
subsetdata(Z, 1:3) # extract first 3 replicates from each data set

NeuralEstimators.subsetparameters — Function

subsetparameters(parameters::M, indices) where {M <: AbstractMatrix}
subsetparameters(parameters::P, indices) where {P <: ParameterConfigurations}

Subset parameters using a collection of indices.

Arrays in parameters::P with last dimension equal in size to the number of parameter configurations, K, are also subsetted (over their last dimension) using indices. All other fields are left unchanged. To modify this default behaviour, overload subsetparameters.

Utility functions

NeuralEstimators.adjacencymatrix — Function

adjacencymatrix(M::Matrix, k::Integer; maxmin::Bool = false)
adjacencymatrix(M::Matrix, r::Float)
adjacencymatrix(M::Matrix, r::Float, k::Integer)

Computes a spatially weighted adjacency matrix from M based on either the k-nearest neighbours of each location; all nodes within a disc of radius r; or, if both r and k are provided, a random set of k neighbours with a disc of radius r.

If maxmin=false (default) the k-nearest neighbours are chosen based on all points in the graph. If maxmin=true, a so-called maxmin ordering is applied, whereby an initial point is selected, and each subsequent point is selected to maximise the minimum distance to those points that have already been selected. Then, the neighbours of each point are defined as the k-nearest neighbours amongst the points that have already appeared in the ordering.

If M is a square matrix, it is treated as a distance matrix; otherwise, it should be an $n$ x d matrix, where $n$ is the number of spatial locations and $d$ is the spatial dimension (typically $d$ = 2). In the latter case, the distance metric is taken to be the Euclidean distance. Note that the maxmin ordering currently requires a set of spatial locations (not a distance matrix).

By convention, we consider a location to neighbour itself and, hence, k-neighbour methods will yield k+1 neighbours for each location. Note that one may use dropzeros!() to remove these self-loops from the constructed adjacency matrix (see below).

Examples

using NeuralEstimators
using Distances
using SparseArrays

n = 100
d = 2
S = rand(n, d)
k = 10
r = 0.1

# Memory efficient constructors (avoids constructing the full distance matrix D)
adjacencymatrix(S, k)
adjacencymatrix(S, k; maxmin = true)
adjacencymatrix(S, r)
adjacencymatrix(S, r, k)

# Construct from full distance matrix D
D = pairwise(Euclidean(), S, S, dims = 1)
adjacencymatrix(D, k)
adjacencymatrix(D, r)
adjacencymatrix(D, r, k)

# Removing self-loops so that a location is not its own neighbour
adjacencymatrix(S, k) |> dropzeros!

NeuralEstimators.containertype — Function

containertype(A::Type)
containertype(::Type{A}) where A <: SubArray
containertype(a::A) where A

Returns the container type of its argument.

If given a SubArray, returns the container type of the parent array.

Examples

a = rand(3, 4)
containertype(a)
containertype(typeof(a))
[containertype(x) for x ∈ eachcol(a)]

NeuralEstimators.encodedata — Function

encodedata(Z::A; fixed_constant::T = zero(T)) where {A <: AbstractArray{Union{Missing, T}, N}} where T, N

For data Z with missing entries, returns an augmented data set (U, W) where W encodes the missingness pattern as an indicator vector and U is the original data Z with missing entries replaced by a fixed_constant.

The indicator vector W is stored in the second-to-last dimension of Z, which should be a singleton. If the second-to-last dimension is not singleton, then two singleton dimensions will be added to the array, and W will be stored in the new second-to-last dimension.

Examples

using NeuralEstimators

# Generate some missing data
Z = rand(16, 16, 1, 1)
Z = removedata(Z, 0.25)	 # remove 25% of the data

# Encode the data
UW = encodedata(Z)

NeuralEstimators.estimateinbatches — Function

estimateinbatches(θ̂, z, θ = nothing; batchsize::Integer = 32, use_gpu::Bool = true, kwargs...)

Apply the estimator θ̂ on minibatches of z (and optionally parameter vectors or other set-level information θ) of size batchsize.

This can prevent memory issues that can occur with large data sets, particularly on the GPU.

Minibatching will only be done if there are multiple data sets in z; this will be inferred by z being a vector, or a tuple whose first element is a vector.

NeuralEstimators.expandgrid — Function

expandgrid(xs, ys)

Same as expand.grid() in R, but currently caters for two dimensions only.

NeuralEstimators.initialise_estimator — Function

initialise_estimator(p::Integer; ...)
initialise_estimator(p::Integer, data_type::String; ...)

Initialise a neural estimator for a statistical model with p unknown parameters.

The estimator is couched in the DeepSets framework (see DeepSet) so that it can be applied to data sets containing an arbitrary number of independent replicates (including the special case of a single replicate).

Note also that the user is free to initialise their neural estimator however they see fit using arbitrary Flux code; see here for Flux's API reference.

Finally, the method with positional argument data_typeis a wrapper that allows one to specify the type of their data (either "unstructured", "gridded", or "irregular_spatial").

Keyword arguments

architecture::String: for unstructured multivariate data, one may use a densely-connected neural network ("DNN"); for data collected over a grid, a convolutional neural network ("CNN"); and for graphical or irregular spatial data, a graphical neural network ("GNN").

d::Integer = 1: for unstructured multivariate data (i.e., when architecture = "DNN"), the dimension of the data (e.g., d = 3 for trivariate data); otherwise, if architecture ∈ ["CNN", "GNN"], the argument d controls the number of input channels (e.g., d = 1 for univariate spatial processes).

estimator_type::String = "point": the type of estimator; either "point" or "interval".

depth = 3: the number of hidden layers; either a single integer or an integer vector of length two specifying the depth of the inner (summary) and outer (inference) network of the DeepSets framework.

width = 32: a single integer or an integer vector of length sum(depth) specifying the width (or number of convolutional filters/channels) in each hidden layer.

activation::Function = relu: the (non-linear) activation function of each hidden layer.

activation_output::Function = identity: the activation function of the output layer.

variance_stabiliser::Union{Nothing, Function} = nothing: a function that will be applied directly to the input, usually to stabilise the variance.

kernel_size = nothing: (applicable only to CNNs) a vector of length depth[1] containing integer tuples of length D, where D is the dimension of the convolution (e.g., D = 2 for two-dimensional convolution).

weight_by_distance::Bool = false: (applicable only to GNNs) flag indicating whether the estimator will weight by spatial distance; if true, a WeightedGraphConv layer is used in the propagation module; otherwise, a regular GraphConv layer is used.

Examples

## DNN, GNN, 1D CNN, and 2D CNN for a statistical model with two parameters:
p = 2
initialise_estimator(p, architecture = "DNN")
initialise_estimator(p, architecture = "GNN")
initialise_estimator(p, architecture = "CNN", kernel_size = [10, 5, 3])
initialise_estimator(p, architecture = "CNN", kernel_size = [(10, 10), (5, 5), (3, 3)])

NeuralEstimators.loadbestweights — Function

loadbestweights(path::String)

Returns the weights of the neural network saved as 'best_network.bson' in the given path.

NeuralEstimators.removedata — Function

removedata(Z::Array, Iᵤ::Vector{Integer})
removedata(Z::Array, p::Union{Float, Vector{Float}}; prevent_complete_missing = true)
removedata(Z::Array, n::Integer; fixed_pattern = false, contiguous_pattern = false, variable_proportion = false)

Replaces elements of Z with missing.

The simplest method accepts a vector of integers Iᵤ that give the specific indices of the data to be removed.

Alterntivaly, there are two methods available to generate data that are missing completely at random (MCAR).

First, a vector p may be given that specifies the proportion of missingness for each element in the response vector. Hence, p should have length equal to the dimension of the response vector. If a single proportion is given, it will be replicated accordingly. If prevent_complete_missing = true, no replicates will contain 100% missingness (note that this can slightly alter the effective values of p).

Second, if an integer n is provided, all replicates will contain n observations after the data are removed. If fixed_pattern = true, the missingness pattern is fixed for all replicates. If contiguous_pattern = true, the data will be removed in a contiguous block. If variable_proportion = true, the proportion of missingness will vary across replicates, with each replicate containing between 1 and n observations after data removal, sampled uniformly (note that variable_proportion overrides fixed_pattern).

The return type is Array{Union{T, Missing}}.

Examples

d = 5           # dimension of each replicate
m = 2000        # number of replicates
Z = rand(d, m)  # simulated data

# Passing a desired proportion of missingness
p = rand(d)
removedata(Z, p)

# Passing a desired final sample size
n = 3  # number of observed elements of each replicate: must have n <= d
removedata(Z, n)

NeuralEstimators.rowwisenorm — Function

rowwisenorm(A)

Computes the row-wise norm of a matrix A.

NeuralEstimators.stackarrays — Function

stackarrays(v::V; merge = true) where {V <: AbstractVector{A}} where {A <: AbstractArray{T, N}} where {T, N}

Stack a vector of arrays v along the last dimension of each array, optionally merging the final dimension of the stacked array.

The arrays must be of the same size for the first N-1 dimensions. However, if merge = true, the size of the final dimension can vary.

Examples

# Vector containing arrays of the same size:
Z = [rand(2, 3, m) for m ∈ (1, 1)];
stackarrays(Z)
stackarrays(Z, merge = false)

# Vector containing arrays with differing final dimension size:
Z = [rand(2, 3, m) for m ∈ (1, 2)];
stackarrays(Z)

NeuralEstimators.vectotril — Function

vectotril(v; strict = false)
vectotriu(v; strict = false)

Converts a vector v of length $d(d+1)÷2$ (a triangular number) into a $d × d$ lower or upper triangular matrix.

If strict = true, the matrix will be strictly lower or upper triangular, that is, a $(d+1) × (d+1)$ triangular matrix with zero diagonal.

Note that the triangular matrix is constructed on the CPU, but the returned matrix will be a GPU array if v is a GPU array. Note also that the return type is not of type Triangular matrix (i.e., the zeros are materialised) since Traingular matrices are not always compatible with other GPU operations.

Examples

using NeuralEstimators

d = 4
n = d*(d+1)÷2
v = collect(range(1, n))
vectotril(v)
vectotriu(v)
vectotril(v; strict = true)
vectotriu(v; strict = true)

Other

NeuralEstimators.EM — Type

EM(simulateconditional::Function, MAP::Function, θ₀ = nothing)

A type that implements the Monte Carlo variant of the expectation-maximisation (EM) algorithm, which at $l$th iteration finds the value of 𝛉 that maximises

\[Σₕᴴ ℓ(𝛉; 𝐙₁, 𝐙₂ˡʰ) + log πᴴ(𝛉),\]

where ℓ(⋅) is the complete-data log-likelihood function, 𝐙 ≡ (𝐙₁', 𝐙₂')' denotes the complete data with 𝐙₁ and 𝐙₂ the observed and missing components, respectively, the replicate 𝐙₂ˡʰ, h = 1, …, H, is sampled from the conditional probability distribution of 𝐙₂ given 𝐙₁ and the previous estimates 𝛉ˡ⁻¹, and πᴴ(⋅) ≡ {π(⋅)}ᴴ is a concentrated version of the original prior density.

Fields

The function simulateconditional should be of the form,

simulateconditional(Z::A, θ; nsims::Integer = 1) where {A <: AbstractArray{Union{Missing, T}}} where T

and the completed-data Z should be returned in whatever form is appropriate to be passed to the MAP estimator as MAP(Z). For example, if the data are gridded and the MAP is a neural MAP estimator based on a CNN architecture, then Z should be returned as a four-dimensional array.

Note that the MAP estimator should return the joint posterior mode; therefore, a neural MAP estimator should be trained under (a surrogate for) the joint 0-1 loss function (see kpowerloss).

The starting values θ₀ should be a vector, which can be provided either during construction of the EM object, or when applying the EM object to data (see below). The starting values given in a function call take precedence over those stored in the object.

Methods

Once constructed, obects of type EM can be applied to data via the methods,

(em::EM)(Z::A, θ₀::Union{Nothing, Vector} = nothing; ...) where {A <: AbstractArray{Union{Missing, T}, N}} where {T, N}
(em::EM)(Z::V, θ₀::Union{Nothing, Vector, Matrix} = nothing; ...) where {V <: AbstractVector{A}} where {A <: AbstractArray{Union{Missing, T}, N}} where {T, N}

where Z is the complete data containing the observed data and Missing values. Note that the second method caters for the case that one has multiple data sets. The keyword arguments are:

niterations::Integer = 50: the maximum number of iterations.

nsims::Integer = 1: the number of conditional replicates used to approximate the conditional expectation.

ξ = nothing: model information needed for conditional simulation (e.g., distance matrices) or in the MAP estimator.

use_ξ_in_simulateconditional::Bool = false: if set to true, the conditional simulator is called as simulateconditional(Z, θ, ξ; nsims = nsims).

use_ξ_in_MAP::Bool = false: if set to true, the MAP estimator is applied to the conditionally-completed data as MAP(Z, ξ).

ϵ = 0.01: tolerance used to assess convergence; The algorithm if the relative change in parameter values from successive iterations is less than ϵ.

return_iterates: if true, the estimate at each iteration of the algorithm is returned; otherwise, only the final estimate is returned.

use_gpu::Bool = true

verbose::Bool = false