Utility functions

Core

These functions can appear during the core workflow, and may need to be overloaded in some applications.

NeuralEstimators.subsetparameters — Function

subsetparameters(parameters::M, indices) where {M <: AbstractMatrix}
subsetparameters(parameters::P, indices) where {P <: ParameterConfigurations}

Subset parameters using a collection of indices.

Arrays in parameters::P with last dimension equal in size to the number of parameter configurations, K, are also subsetted (over their last dimension) using indices. All other fields are left unchanged. To modify this default behaviour, overload subsetparameters.

NeuralEstimators.numberreplicates — Function

numberofreplicates(Z)

Generic function that returns the number of replicates in a given object. Default implementations are provided for commonly used data formats, namely, data stored as an Array or as a GNNGraph.

NeuralEstimators.subsetdata — Function

Generic function for subsetting replicates from a data set. Default methods are:

subsetdata(Z::A, m) where {A <: AbstractArray{T, N}} where {T, N}
subsetdata(Z::G, m) where {G <: AbstractGraph}

Note that subsetdata is slow for graphical data, and one should consider using a method of train that does not require the data to be subsetted. Use numberreplicates to check that the training and validation data sets are equally replicated, which prevents the invocation of subsetdata. Note also that subsetdata only applies to vectors of batched graphs.

If the user is working with data that is not covered by the default methods, simply overload subsetdata with the appropriate type for Z.

Examples

using NeuralEstimators
using GraphNeuralNetworks
using Flux: batch

n = 5  # number of observations in each realisation
m = 6  # number of replicates for each parameter vector
d = 1  # dimension of the response variable
K = 2  # number of parameter vectors

# Array data
Z = [rand(n, d, m) for k ∈ 1:K]
subsetdata(Z, 1:3) # extract first 3 replicates for each parameter vector

# Graphical data
e = 8 # number of edges
Z = [batch([rand_graph(n, e, ndata = rand(d, n)) for _ ∈ 1:m]) for k ∈ 1:K]
subsetdata(Z, 1:3) # extract first 3 replicates for each parameter vector

Miscellaneous

NeuralEstimators.loadbestweights — Function

loadbestweights(path::String)

Given a path to a training run containing neural networks saved with names "network_epochx.bson" and an object saved as "loss_per_epoch.bson", returns the weights of the best network (measured by validation loss).

NeuralEstimators.stackarrays — Function

stackarrays(v::V; merge = true) where {V <: AbstractVector{A}} where {A <: AbstractArray{T, N}} where {T, N}

Stack a vector of arrays v along the last dimension of each array, optionally merging the final dimension of the stacked array.

The arrays must be of the same size for the first N-1 dimensions. However, if merge = true, the size of the final dimension can vary.

Examples

# Vector containing arrays of the same size:
Z = [rand(2, 3, m) for m ∈ (1, 1)];
stackarrays(Z)
stackarrays(Z, merge = false)

# Vector containing arrays with differing final dimension size:
Z = [rand(2, 3, m) for m ∈ (1, 2)];
stackarrays(Z)

NeuralEstimators.expandgrid — Function

expandgrid(xs, ys)

Same as expand.grid() in R, but currently caters for two dimensions only.