Core functions

This page documents the functions that are central to the workflow of NeuralEstimators. Its organisation reflects the order in which these functions appear in a standard implementation; that is, from storing parameters sampled from the prior distribution, to uncertainty quantification of the final estimates via bootstrapping.

Storing parameters

Parameters sampled from the prior distribution $\Omega(\cdot)$ may be stored i) as a $p \times K$ matrix, where $p$ is the number of parameters in the model and $K$ is the number of parameter vectors sampled from the prior distribution, or ii) in a user-defined subtype of the abstract type ParameterConfigurations. See Storing expensive intermediate objects for data simulation for further discussion.

NeuralEstimators.ParameterConfigurations — Type

ParameterConfigurations

An abstract supertype for user-defined types that store parameters and any intermediate objects needed for data simulation with simulate.

The user-defined type must have a field θ that stores the $p$ × $K$ matrix of parameters, where $p$ is the number of parameters in the model and $K$ is the number of parameter vectors sampled from the prior distribution. There are no other restrictions.

Examples

struct P <: ParameterConfigurations
	θ
	# ...
end

NeuralEstimators.subsetparameters — Function

subsetparameters(parameters::M, indices) where {M <: AbstractMatrix}
subsetparameters(parameters::P, indices) where {P <: ParameterConfigurations}

Subset parameters using a collection of indices.

Arrays in parameters::P with last dimension equal in size to the number of parameter configurations, K, are also subsetted (over their last dimension) using indices. All other fields are left unchanged. To modify this default behaviour, redefine subsetparameters after running import NeuralEstimators: subsetparameters.

Simulating data

NeuralEstimators facilitates neural estimation for arbitrary statistical models by having the user implicitly define their model either by providing simulated data, or by defining a function for data simulation. If the latter option is chosen, the user must provide a method simulate(parameters, m), which returns simulated data from a set of parameters, with m the sample size of these simulated data.

Irrespective of their source, the simulated data must be stored as a subtype of AbstractVector{AbstractArray}, where each array stores m independent replicates simulated from one parameter vector in parameters, and these replicates must stored in the final dimension of each array.

NeuralEstimators.simulate — Function

Generic function that the user may provide methods for in order to implicitly define their statistical model.

NeuralEstimators.simulate — Method

simulate(parameters, m::Integer, J::Integer)

Simulates J sets of m independent replicates for each parameter vector in parameters by calling simulate(parameters, m) a total of J times.

Representations for neural estimators

Although the user is free to construct their neural estimator however they see fit, NeuralEstimators provides several useful representations described below. Note that if the user wishes to use an alternative representation, for compatibility with NeuralEstimators, simply ensure that the estimator processes data stored as subtypes of AbstractVector{AbstractArray}, as discussed in DeepSet (see also its source code).

Deep Set

NeuralEstimators.DeepSet — Type

DeepSet(ψ, ϕ, a)
DeepSet(ψ, ϕ; a::String = "mean")

A Deep Set neural estimator,

\[θ̂(𝐙) ≡ ϕ(a(\{𝐙ᵢ : i = 1, …, m\})),\]

where 𝐙 ≡ (𝐙₁', …, 𝐙ₘ')' are independent and identically distributed (iid) realisations from the model under a single parameter vector 𝛉, ψ and ϕ are neural networks, and a is a permutation-invariant aggregation function. Note that ψ and ϕ depend on trainable parameters, but we omit this dependence for notational convenience.

Although the above defintion of a neural estimator is with respect to a single data set 𝐙, DeepSet estimators instead act on sets of data sets, stored as Vectors of Arrays, where each array corresponds to one set of iid realisations from the model. The last dimension of each array stores the realisations; for example, if 𝐙 is a 3-dimensional array, then 𝐙[:, :, 1] contains the first realisation, 𝐙[:, :, 2] contains the second realisation, and so on.

The neural networks ψ and ϕ and typically Flux neural networks. The function a must act on an Array and, since it aggregates the iid realisations, it must aggregate over the last dimension of the array.

There are two DeepSet constructors. The first constructor treats a as a keyword argument of type String, which can take values "mean" (default), "sum", or "logsumexp" function. The second constructor treats a as a positional argument of type Function, and this constructor allows the user to provide a custom aggregation function.

Examples

n = 10 # observations in each realisation
p = 5  # number of parameters in the statistical model
w = 32 # width of each layer
ψ = Chain(Dense(n, w, relu), Dense(w, w, relu));
ϕ = Chain(Dense(w, w, relu), Dense(w, p));
θ̂ = DeepSet(ψ, ϕ)

# Apply the estimator to a single set of m=3 realisations:
Z = [rand(n, 1, 3)];
θ̂(Z)

# Apply the estimator to two sets each containing m=3 realisations:
Z = [rand(n, 1, m) for m ∈ (3, 3)];
θ̂(Z)

# Apply the estimator to two sets containing m=3 and m=4 realisations, respectively:
Z = [rand(n, 1, m) for m ∈ (3, 4)];
θ̂(Z)

Piecewise estimators

NeuralEstimators.PiecewiseEstimator — Type

PiecewiseEstimator(estimators, mchange)

Creates a piecewise estimator from a collection of estimators, based on the collection of sample-size changepoints, mchange, which should contain one element fewer than the number of estimators.

Examples

Suppose that we've trained θ̂₁ for small sample sizes (e.g., m ≤ 30) and θ̂₂ for moderate-to-large sample sizes (e.g., m > 30). Then we can construct a piecewise estimator with a sample-size changepoint of 30, which dispatches θ̂₁ if m ≤ 30 and θ̂₂ if m > 30:

n = 2
p = 3
w = 8

ψ₁ = Chain(Dense(n, w, relu), Dense(w, w, relu));
ϕ₁ = Chain(Dense(w, w, relu), Dense(w, p));
θ̂₁ = DeepSet(ψ₁, ϕ₁)

ψ₂ = Chain(Dense(n, w, relu), Dense(w, w, relu), Dense(w, w, relu));
ϕ₂ = Chain(Dense(w, w, relu), Dense(w, w, relu), Dense(w, p));
θ̂₂ = DeepSet(ψ₂, ϕ₂)

θ̂ = PiecewiseEstimator([θ̂₁, θ̂₂], [30])
Z = [rand(n, 1, m) for m ∈ (10, 50)]
θ̂(Z)

Training

NeuralEstimators.train — Function

Generic function for training a neural estimator.

The methods are designed to cater for different forms of "on-the-fly simulation" (see the online documentation). In all methods, the validation data are held fixed so that the validation risk function, which is used to monitor the performance of the estimator during training, is not subject to noise.

Note that train is a mutating function, but the suffix ! is omitted to avoid clashes with the Flux function, train!.

Keyword arguments

Arguments common to all methods:

loss = mae: the loss function, which should return the average loss when applied to multiple replicates.

epochs::Integer = 100

batchsize::Integer = 32

optimiser = ADAM(1e-4)

savepath::String = "": path to save the trained θ̂ and other information; if savepath is an empty string (default), nothing is saved.

stopping_epochs::Integer = 10: cease training if the risk doesn't improve in stopping_epochs epochs.

use_gpu::Bool = true

verbose::Bool = true

Arguments common to train(θ̂, P) and train(θ̂, θ_train, θ_val):

m: sample sizes (either an Integer or a collection of Integers).

epochs_per_Z_refresh::Integer = 1: how often to refresh the training data.

simulate_just_in_time::Bool = false: should we simulate the data "just-in-time"?

Arguments unique to train(θ̂, P):

K::Integer = 10_000: number of parameter vectors in the training set; the size of the validation set is K ÷ 5.

epochs_per_θ_refresh::Integer = 1: how often to refresh the training parameters; this must be a multiple of epochs_per_Z_refresh.

ξ = nothing: invariant model information; if ξ is provided, the constructor P is called as P(K, ξ).

NeuralEstimators.train — Method

train(θ̂, P; <keyword args>)

Train the neural estimator θ̂ by providing a constructor, P, where P is a subtype of AbstractMatrix or ParameterConfigurations, to automatically sample the sets of training and validation parameters.

NeuralEstimators.train — Method

train(θ̂, θ_train::P, θ_val::P; <keyword args>)

Train the neural estimator θ̂ by providing the training and validation parameter sets explicitly as θ_train and θ_val, which are both held fixed during training.

NeuralEstimators.train — Method

train(θ̂, θ_train::P, θ_val::P, Z_train::T, Z_val::T; <keyword args>)

Train the neural estimator θ̂ by providing the training and validation parameter sets, θ_train and θ_val, and the training and validation data sets, Z_train and Z_val, all of which are held fixed during training.

The sample size argument m is inferred from Z_val. The training data Z_train can contain M replicates, where M is a multiple of m; the training data will then be recycled to imitate on-the-fly simulation. For example, if M = 50 and m = 10, epoch 1 uses the first 10 replicates, epoch 2 uses the next 10 replicates, and so on, until epoch 6 again uses the first 10 replicates.

Note that the elements of Z_train and Z_val should each be equally replicated; that is, the size of the last dimension in each array in Z_train should be constant, and similarly for Z_val (although these constants can differ, as discussed above).

NeuralEstimators.train — Method

train(θ̂, θ_train::P, θ_val::P, Z_train::T, Z_val::T, M; <keyword args>)

Train several neural estimators, each with the architecture given by θ̂, using the sample sizes given by the vector of integers, M.

Each neural estimator is pre-trained with the neural estimator trained for the previous sample size. That is, if M = [m₁, m₂], with m₂ > m₁, the neural estimator for sample size m₂ is pre-trainined with the (trained) neural estimator for sample size m₁. By pre-training a series of neural estimators with progressively larger sample sizes, most of the learning is done with small, computationally cheap sample sizes. Hence, this approach can be beneficial even if one is only interested in estimation for a single, large sample.

This method is a wrapper for train(θ̂, θ_train::P, θ_val::P, Z_train::T, Z_val::T) and, hence, it inherits its keyword arguments. Further, certain keyword arguments can be given as vectors. For instance, if we are training two neural estimators, we can use a different number of epochs by providing epochs = [e₁, e₂]. Other arguments that allow vectors are batchsize, stopping_epochs, and optimiser.

Assessing a neural estimator

NeuralEstimators.assess — Function

assess(estimators, parameters, Z; <keyword args>)
assess(estimators, parameters; <keyword args>)

Using a collection of estimators, compute estimates from data simulated from a set of parameters.

One may either provide simulated data Z as a Vector{Vector{Array}}, or overload simulate with a method simulate(parameters, m::Integer) and use the keyword argument m to specify the sample sizes to use during assessment.

Keyword arguments

m::Vector{Integer}: sample sizes to estimate from.

estimator_names::Vector{String}: names of the estimators (sensible default values provided).

parameter_names::Vector{String}: names of the parameters (sensible default values provided).

J::Integer = 1: the number of times to replicate each parameter in parameters.

ξ = nothing: invariant model information.

use_ξ = false: a Bool or a collection of Bool objects with length equal to the number of estimators. Specifies whether or not the estimator uses the invariant model information, ξ: If it does, the estimator will be applied as estimator(Z, ξ).

use_gpu = true: a Bool or a collection of Bool objects with length equal to the number of estimators.

verbose::Bool = true

NeuralEstimators.Assessment — Type

Assessment(θandθ̂, runtime)

An object for storing the result of calling assess(), containing a DataFrame of true parameters θ and corresponding estimates θ̂, and a DataFrame containing the runtime for each estimator.

Bootstrapping

Note that all bootstrapping functions are currently implemented for a single parameter configuration only.

NeuralEstimators.parametricbootstrap — Function

parametricbootstrap(θ̂, parameters::P, m::Integer; B::Integer = 100, use_gpu::Bool = true) where {P <: ParameterConfigurations}

Returns B parameteric bootstrap samples of an estimator θ̂ as a p × B matrix, where p is the number of parameters in the statistical model, based on data sets of size m simulated parameter configurations, parameters.

This function requires the user to have defined a method simulate(parameters, m::Integer).

NeuralEstimators.nonparametricbootstrap — Function

nonparametricbootstrap(θ̂, Z::AbstractArray{T, N}; B::Integer = 100, use_gpu::Bool = true)
nonparametricbootstrap(θ̂, Z::AbstractArray{T, N}, blocks; B::Integer = 100, use_gpu::Bool = true)

Returns B non-parametric bootstrap samples of an estimator θ̂ as a p × B matrix, where p is the number of parameters in the statistical model.

The argument blocks caters for block bootstrapping, and should be an integer vector specifying the block for each replicate. For example, if we have 5 replicates with the first two replicates corresponding to block 1 and the remaining replicates corresponding to block 2, then blocks should be [1, 1, 2, 2, 2]. The resampling algorithm tries to produce resampled data sets of a similar size to the original data, but this can only be achieved exactly if the blocks are the same length.