Core functions

Deep Set representation

NeuralEstimators.DeepSet — Type

DeepSet(ψ, ϕ, agg)

Implementation of the Deep Set framework, where ψ and ϕ are neural networks (e.g., Flux networks) and agg is a symmetric function that pools data over the last dimension (the replicates/batch dimension) of an array.

DeepSet objects are applied to AbstractVectors of AbstractArrays, where each array is associated with one parameter vector.

Examples

n = 10 # observations in each realisation
p = 5  # number of parameters in the statistical model 
w = 32 # width of each layer
ψ = Chain(Dense(n, w, relu), Dense(w, w, relu));
ϕ = Chain(Dense(w, w, relu), Dense(w, p));
agg(X) = sum(X, dims = ndims(X))
θ̂  = DeepSet(ψ, ϕ, agg)

# A single set of m=3 realisations:
Z = [rand(n, 1, 3)];
θ̂ (Z)

# Two sets each containing m=3 realisations:
Z = [rand(n, 1, m) for m ∈ (3, 3)];
θ̂ (Z)

# Two sets respectivaly containing m=3 and m=4 realisations:
Z = [rand(n, 1, m) for m ∈ (3, 4)];
θ̂ (Z)

NeuralEstimators.DeepSet — Method

DeepSet(ψ, ϕ; aggregation::String = "mean")

Convenient constructor for a DeepSet object with agg equal to the "mean", "sum", or "log-sum-exp" function.

Simulation

NeuralEstimators.simulate — Function

simulate(parameters::P, m::Integer, num_rep::Integer) where {P <: ParameterConfigurations}

Generic method that simulates m independent replicates for each parameter configuration (by internally calling simulate(parameters, m)), repeated a total of num_rep times.

See also Data simulation.

Training

There are two training methods. For both methods, the validation parameters and validation data are held fixed so that the validation risk is interpretable. There are a number of practical considerations to keep in mind: In particular, see Balancing time and memory complexity.

NeuralEstimators.train — Function

train(θ̂, ξ, P; <keyword args>) where {P <: ParameterConfigurations}

Train the neural estimator θ̂ by providing the objects ξ needed for the constructor P to automatically sample the set of training and validation parameters, which may be refreshed periodically throughout training via the keyword argument epochs_per_θ_refresh.

Keyword arguments common to both train methods:

m: sample sizes (either an Integer or a collection of Integers).

batchsize::Integer = 32

epochs::Integer = 100: the maximum number of epochs used during training.

epochs_per_Z_refresh::Integer = 1: how often to refresh the training data.

loss = mae: the loss function, which should return an average loss when applied to multiple replicates.

optimiser = ADAM(1e-4)

savepath::String = "runs/": path to save the trained θ̂ and other information.

stopping_epochs::Integer = 10: halt training if the risk doesn't improve in stopping_epochs epochs.

use_gpu::Bool = true

Simulator keyword arguments only:

configs_per_epoch::Integer = 10_000: how many parameters constitute a single epoch.

epochs_per_θ_refresh::Integer = 1: how often to refresh the training parameters; this must be a multiple of epochs_per_Z_refresh.

train(θ̂, θ_train::P, θ_val::P; <keyword args>) where {P <: ParameterConfigurations}

Train the neural estimator θ̂ by providing the training and validation sets explicitly as θ_train and θ_val, which are both held fixed during training.

Estimation

NeuralEstimators.estimate — Function

estimate(estimators, parameters::P, m; <keyword args>) where {P <: ParameterConfigurations}

Using a collection of estimators, compute estimates from data simulated from a set of parameters.

estimate() requires the user to have defined a method simulate(parameters, m::Integer).

Keyword arguments

m::Vector{Integer} where I <: Integer: sample sizes to estimate from.

estimator_names::Vector{String}: estimator names used when combining estimates into a DataFrame (e.g., ["NeuralEstimator", "BayesEstimator", "MLE"]), with sensible default values provided.

parameter_names::Vector{String}: parameter names used when combining estimates into a DataFrame (e.g., ["μ", "σ"]), with sensible default values provided.

num_rep::Integer = 1: the number of times to replicate each parameter in parameters to reduce the effect of sample variability when assessing the estimators.

use_gpu = true: should be a Bool or a collection of Bool objects with length equal to the number of estimators.

Bootstrapping

Note that all bootstrapping functions are currently implemented for a single parameter configuration only.

NeuralEstimators.parametricbootstrap — Function

parametricbootstrap(θ̂, parameters::P, m::Integer; B::Integer = 100, use_gpu::Bool = true) where {P <: ParameterConfigurations}

Returns B parameteric bootstrap samples of an estimator θ̂ as a p × B matrix, where p is the number of parameters in the statistical model, based on simulated data sets of size m.

This function requires a method simulate(parameters::P, m::Integer).

NeuralEstimators.nonparametricbootstrap — Function

nonparametricbootstrap(θ̂, Z; B::Integer = 100, use_gpu::Bool = true)
nonparametricbootstrap(θ̂, Z, blocks; B::Integer = 100, use_gpu::Bool = true)

Returns B non-parametric bootstrap samples of an estimator θ̂ as a p × B matrix, where p is the number of parameters in the statistical model.

The argument blocks caters for block bootstrapping, and should be an integer vector specifying the block for each replicate. For example, if we have 5 replicates with the first two replicates corresponding to block 1 and the remaining replicates corresponding to block 2, then blocks should be [1, 1, 2, 2, 2]. The resampling algorithm tries to produce resampled data sets of a similar size to the original data, but this can only be achieved exactly if the blocks are the same length.