Model-specific functions

Data simulators

The philosophy of NeuralEstimators is to cater for arbitrary statistical models by having the user define their statistical model implicitly, either by providing data simulated from the model or by defining a function for data simulation. However, the following functions, which were developed for the main manuscript, have been included as they may be of use to others, and their source code provide an example for how a user could formulate code for their own statistical model. If you've developed similar functions that you think may be of use to others, please get in touch or make a pull request.

NeuralEstimators.simulategaussianprocess — Function

simulategaussianprocess(L::AbstractArray{T, 2}, σ::T, m::Integer)
simulategaussianprocess(L::AbstractArray{T, 2})

Simulates m realisations from a Gau(0, 𝚺 + σ²𝐈) distribution, where 𝚺 ≡ LL'.

If σ and m are omitted, a single field without nugget variance is returned.

NeuralEstimators.simulateschlather — Function

simulateschlather(L::AbstractArray{T, 2}; C = 3.5)
simulateschlather(L::AbstractArray{T, 2}, m::Integer; C = 3.5)

Simulates from Schlather's max-stable model.

Commonly-used low-level functions

These low-level functions may be of use for various models.

NeuralEstimators.matern — Function

matern(h, ρ, ν, σ² = 1)

For two points separated by h units, compute the Matérn covariance function with range ρ, smoothness ν, and marginal variance σ².

We use the parametrisation $C(\mathbf{h}) = \sigma^2 \frac{2^{1 - \nu}}{\Gamma(\nu)} \left(\frac{\|\mathbf{h}\|}{\rho}\right) K_\nu \left(\frac{\|\mathbf{h}\|}{\rho}\right)$, where $\Gamma(\cdot)$ is the gamma function, and $K_\nu(\cdot)$ is the modified Bessel function of the second kind of order $\nu$. This parameterisation is the same as used by the R package fields, but differs to the parametrisation given by Wikipedia.

Note that the Julia functions for $\Gamma(\cdot)$ and $K_\nu(\cdot)$, respectively gamma() and besselk(), do not work on the GPU and, hence, nor does matern().

NeuralEstimators.maternchols — Function

maternchols(D, ρ, ν)

Given a distance matrix D, computes the covariance matrix under the Matérn covariance function with range ρ and smoothness ν, and returns the Cholesky factor of this covariance matrix.

Providing vectors for ρ and ν will yield a three-dimensional array of Cholesky factors. Similarly for a vector of matrices D.

NeuralEstimators.incgamma — Function

incgamma(a::T, x::T; upper::Bool, reg::Bool) where {T <: AbstractFloat}

For positive parameter a and positive integration limit x, computes the incomplete gamma function, as described by the Wikipedia article.

Keyword arguments:

upper::Bool: if true, the upper incomplete gamma function is returned; otherwise, the lower version is returned.

reg::Bool: if true, the regularized incomplete gamma function is returned; otherwise, the unregularized version is returned.

Density functions

Density functions are not needed in the workflow of NeuralEstimators. However, as part of a series of comparison studies between neural estimators and likelihood-based estimators given in the manuscript, we have developed the following density functions, and we include them in NeuralEstimators to cater for the possibility that they may be of use in future comparison studies.

NeuralEstimators.gaussiandensity — Function

gaussiandensity(y::A, L; logdensity::Bool = true) where {A <: AbstractArray{T, 1}} where T
gaussiandensity(y::A, Σ; logdensity::Bool = true) where {A <: AbstractArray{T, N}} where {T, N}

Efficiently computes the density function for y ~ 𝑁(0, Σ), with L the lower Cholesky factor of the covariance matrix Σ.

The method gaussiandensity(y::A, Σ) assumes that the last dimension of y corresponds to the indepdenent-replicates dimension, and it exploits the fact that we need to compute the Cholesky factor L for these independent replicates once only.

NeuralEstimators.schlatherbivariatedensity — Function

schlatherbivariatedensity(z₁, z₂, ψ; logdensity::Bool = true)

The bivariate density function for Schlather's max-stable model, as given in Raphaël Huser's PhD thesis (pg. 231-232) and in the supplementary material of the manuscript.